refactor(tests): BACK-44 moved tests to a separate folder
