+++ Dockerfile.dev
@@ -26,6 +26,6 @@ HEALTHCHECK --interval=5s --timeout=5s --retries=3 \
     CMD curl -f http://localhost:8081/health || exit 1
 
 # Run the uvicorn command, telling it to use the app object imported from app.main.
-CMD ["uvicorn", "app.main:app", "--host", "0.0.0.0", "--port", "8081", "--reload", "--reload-exclude", "app/test_main.py"]
+CMD ["uvicorn", "app.main:app", "--host", "0.0.0.0", "--port", "8081", "--reload", "--reload-exclude", "tests/*.py"]
 
-# CMD ["uvicorn", "app.main:app", "--proxy-headers", "--host", "0.0.0.0", "--port", "8081" "--reload", "--reload-exclude", "app/test_main.py"]
+# CMD ["uvicorn", "app.main:app", "--proxy-headers", "--host", "0.0.0.0", "--port", "8081" "--reload", "--reload-exclude", "tests/*.py"]
