--- app/test_main.py
@@ -1,4155 +1,0 @@
-# External
-from fastapi import status
-from fastapi.testclient import TestClient
-
-# App
-from .main import app
-
-
-client = TestClient(app)
-
-
-# region root
-
-
-def test_read_root():
-    response = client.get("/")
-    assert response.status_code == status.HTTP_200_OK
-    assert response.json() == {"we_are": "SlickTelemetry"}
-
-
-# endregion root
-
-
-# region healthcheck
-
-
-def test_healthcheck():
-    response = client.get("/health")
-    assert response.status_code == status.HTTP_200_OK
-    assert response.json() == {"status": "OK"}
-
-
-# endregion healtcheck
-
-
-# region schedule
-
-# region schedule - good inputs
-
-
-def test_get_schedule():
-    response = client.get("/schedule")
-    assert response.status_code == status.HTTP_200_OK
-    assert response.json() == {
-        "year": 2023,
-        "EventSchedule": [
-            {
-                "RoundNumber": 1,
-                "Country": "Bahrain",
-                "Location": "Sakhir",
-                "OfficialEventName": "FORMULA 1 GULF AIR BAHRAIN GRAND PRIX 2023",
-                "EventDate": "2023-03-05",
-                "EventName": "Bahrain Grand Prix",
-                "EventFormat": "conventional",
-                "Session1": "Practice 1",
-                "Session1Date": "2023-03-03 14:30:00+03:00",
-                "Session1DateUtc": "2023-03-03 11:30:00",
-                "Session2": "Practice 2",
-                "Session2Date": "2023-03-03 18:00:00+03:00",
-                "Session2DateUtc": "2023-03-03 15:00:00",
-                "Session3": "Practice 3",
-                "Session3Date": "2023-03-04 14:30:00+03:00",
-                "Session3DateUtc": "2023-03-04 11:30:00",
-                "Session4": "Qualifying",
-                "Session4Date": "2023-03-04 18:00:00+03:00",
-                "Session4DateUtc": "2023-03-04 15:00:00",
-                "Session5": "Race",
-                "Session5Date": "2023-03-05 18:00:00+03:00",
-                "Session5DateUtc": "2023-03-05 15:00:00",
-                "F1ApiSupport": True,
-            },
-            {
-                "RoundNumber": 2,
-                "Country": "Saudi Arabia",
-                "Location": "Jeddah",
-                "OfficialEventName": "FORMULA 1 STC SAUDI ARABIAN GRAND PRIX 2023",
-                "EventDate": "2023-03-19",
-                "EventName": "Saudi Arabian Grand Prix",
-                "EventFormat": "conventional",
-                "Session1": "Practice 1",
-                "Session1Date": "2023-03-17 16:30:00+03:00",
-                "Session1DateUtc": "2023-03-17 13:30:00",
-                "Session2": "Practice 2",
-                "Session2Date": "2023-03-17 20:00:00+03:00",
-                "Session2DateUtc": "2023-03-17 17:00:00",
-                "Session3": "Practice 3",
-                "Session3Date": "2023-03-18 16:30:00+03:00",
-                "Session3DateUtc": "2023-03-18 13:30:00",
-                "Session4": "Qualifying",
-                "Session4Date": "2023-03-18 20:00:00+03:00",
-                "Session4DateUtc": "2023-03-18 17:00:00",
-                "Session5": "Race",
-                "Session5Date": "2023-03-19 20:00:00+03:00",
-                "Session5DateUtc": "2023-03-19 17:00:00",
-                "F1ApiSupport": True,
-            },
-            {
-                "RoundNumber": 3,
-                "Country": "Australia",
-                "Location": "Melbourne",
-                "OfficialEventName": "FORMULA 1 ROLEX AUSTRALIAN GRAND PRIX 2023",
-                "EventDate": "2023-04-02",
-                "EventName": "Australian Grand Prix",
-                "EventFormat": "conventional",
-                "Session1": "Practice 1",
-                "Session1Date": "2023-03-31 12:30:00+10:00",
-                "Session1DateUtc": "2023-03-31 02:30:00",
-                "Session2": "Practice 2",
-                "Session2Date": "2023-03-31 16:00:00+10:00",
-                "Session2DateUtc": "2023-03-31 06:00:00",
-                "Session3": "Practice 3",
-                "Session3Date": "2023-04-01 12:30:00+10:00",
-                "Session3DateUtc": "2023-04-01 02:30:00",
-                "Session4": "Qualifying",
-                "Session4Date": "2023-04-01 16:00:00+10:00",
-                "Session4DateUtc": "2023-04-01 06:00:00",
-                "Session5": "Race",
-                "Session5Date": "2023-04-02 15:00:00+10:00",
-                "Session5DateUtc": "2023-04-02 05:00:00",
-                "F1ApiSupport": True,
-            },
-            {
-                "RoundNumber": 4,
-                "Country": "Azerbaijan",
-                "Location": "Baku",
-                "OfficialEventName": "FORMULA 1 AZERBAIJAN GRAND PRIX 2023",
-                "EventDate": "2023-04-30",
-                "EventName": "Azerbaijan Grand Prix",
-                "EventFormat": "sprint_shootout",
-                "Session1": "Practice 1",
-                "Session1Date": "2023-04-28 13:30:00+04:00",
-                "Session1DateUtc": "2023-04-28 09:30:00",
-                "Session2": "Qualifying",
-                "Session2Date": "2023-04-28 17:00:00+04:00",
-                "Session2DateUtc": "2023-04-28 13:00:00",
-                "Session3": "Sprint Shootout",
-                "Session3Date": "2023-04-29 12:30:00+04:00",
-                "Session3DateUtc": "2023-04-29 08:30:00",
-                "Session4": "Sprint",
-                "Session4Date": "2023-04-29 17:30:00+04:00",
-                "Session4DateUtc": "2023-04-29 13:30:00",
-                "Session5": "Race",
-                "Session5Date": "2023-04-30 15:00:00+04:00",
-                "Session5DateUtc": "2023-04-30 11:00:00",
-                "F1ApiSupport": True,
-            },
-            {
-                "RoundNumber": 5,
-                "Country": "United States",
-                "Location": "Miami",
-                "OfficialEventName": "FORMULA 1 CRYPTO.COM MIAMI GRAND PRIX 2023",
-                "EventDate": "2023-05-07",
-                "EventName": "Miami Grand Prix",
-                "EventFormat": "conventional",
-                "Session1": "Practice 1",
-                "Session1Date": "2023-05-05 14:00:00-04:00",
-                "Session1DateUtc": "2023-05-05 18:00:00",
-                "Session2": "Practice 2",
-                "Session2Date": "2023-05-05 17:30:00-04:00",
-                "Session2DateUtc": "2023-05-05 21:30:00",
-                "Session3": "Practice 3",
-                "Session3Date": "2023-05-06 12:30:00-04:00",
-                "Session3DateUtc": "2023-05-06 16:30:00",
-                "Session4": "Qualifying",
-                "Session4Date": "2023-05-06 16:00:00-04:00",
-                "Session4DateUtc": "2023-05-06 20:00:00",
-                "Session5": "Race",
-                "Session5Date": "2023-05-07 15:30:00-04:00",
-                "Session5DateUtc": "2023-05-07 19:30:00",
-                "F1ApiSupport": True,
-            },
-            {
-                "RoundNumber": 6,
-                "Country": "Monaco",
-                "Location": "Monaco",
-                "OfficialEventName": "FORMULA 1 GRAND PRIX DE MONACO 2023",
-                "EventDate": "2023-05-28",
-                "EventName": "Monaco Grand Prix",
-                "EventFormat": "conventional",
-                "Session1": "Practice 1",
-                "Session1Date": "2023-05-26 13:30:00+02:00",
-                "Session1DateUtc": "2023-05-26 11:30:00",
-                "Session2": "Practice 2",
-                "Session2Date": "2023-05-26 17:00:00+02:00",
-                "Session2DateUtc": "2023-05-26 15:00:00",
-                "Session3": "Practice 3",
-                "Session3Date": "2023-05-27 12:30:00+02:00",
-                "Session3DateUtc": "2023-05-27 10:30:00",
-                "Session4": "Qualifying",
-                "Session4Date": "2023-05-27 16:00:00+02:00",
-                "Session4DateUtc": "2023-05-27 14:00:00",
-                "Session5": "Race",
-                "Session5Date": "2023-05-28 15:00:00+02:00",
-                "Session5DateUtc": "2023-05-28 13:00:00",
-                "F1ApiSupport": True,
-            },
-            {
-                "RoundNumber": 7,
-                "Country": "Spain",
-                "Location": "Barcelona",
-                "OfficialEventName": "FORMULA 1 AWS GRAN PREMIO DE ESPAÑA 2023",
-                "EventDate": "2023-06-04",
-                "EventName": "Spanish Grand Prix",
-                "EventFormat": "conventional",
-                "Session1": "Practice 1",
-                "Session1Date": "2023-06-02 13:30:00+02:00",
-                "Session1DateUtc": "2023-06-02 11:30:00",
-                "Session2": "Practice 2",
-                "Session2Date": "2023-06-02 17:00:00+02:00",
-                "Session2DateUtc": "2023-06-02 15:00:00",
-                "Session3": "Practice 3",
-                "Session3Date": "2023-06-03 12:30:00+02:00",
-                "Session3DateUtc": "2023-06-03 10:30:00",
-                "Session4": "Qualifying",
-                "Session4Date": "2023-06-03 16:00:00+02:00",
-                "Session4DateUtc": "2023-06-03 14:00:00",
-                "Session5": "Race",
-                "Session5Date": "2023-06-04 15:00:00+02:00",
-                "Session5DateUtc": "2023-06-04 13:00:00",
-                "F1ApiSupport": True,
-            },
-            {
-                "RoundNumber": 8,
-                "Country": "Canada",
-                "Location": "Montréal",
-                "OfficialEventName": "FORMULA 1 PIRELLI GRAND PRIX DU CANADA 2023",
-                "EventDate": "2023-06-18",
-                "EventName": "Canadian Grand Prix",
-                "EventFormat": "conventional",
-                "Session1": "Practice 1",
-                "Session1Date": "2023-06-16 13:30:00-04:00",
-                "Session1DateUtc": "2023-06-16 17:30:00",
-                "Session2": "Practice 2",
-                "Session2Date": "2023-06-16 16:30:00-04:00",
-                "Session2DateUtc": "2023-06-16 20:30:00",
-                "Session3": "Practice 3",
-                "Session3Date": "2023-06-17 12:30:00-04:00",
-                "Session3DateUtc": "2023-06-17 16:30:00",
-                "Session4": "Qualifying",
-                "Session4Date": "2023-06-17 16:00:00-04:00",
-                "Session4DateUtc": "2023-06-17 20:00:00",
-                "Session5": "Race",
-                "Session5Date": "2023-06-18 14:00:00-04:00",
-                "Session5DateUtc": "2023-06-18 18:00:00",
-                "F1ApiSupport": True,
-            },
-            {
-                "RoundNumber": 9,
-                "Country": "Austria",
-                "Location": "Spielberg",
-                "OfficialEventName": "FORMULA 1 ROLEX GROSSER PREIS VON ÖSTERREICH 2023",
-                "EventDate": "2023-07-02",
-                "EventName": "Austrian Grand Prix",
-                "EventFormat": "sprint_shootout",
-                "Session1": "Practice 1",
-                "Session1Date": "2023-06-30 13:30:00+02:00",
-                "Session1DateUtc": "2023-06-30 11:30:00",
-                "Session2": "Qualifying",
-                "Session2Date": "2023-06-30 17:00:00+02:00",
-                "Session2DateUtc": "2023-06-30 15:00:00",
-                "Session3": "Sprint Shootout",
-                "Session3Date": "2023-07-01 12:00:00+02:00",
-                "Session3DateUtc": "2023-07-01 10:00:00",
-                "Session4": "Sprint",
-                "Session4Date": "2023-07-01 16:30:00+02:00",
-                "Session4DateUtc": "2023-07-01 14:30:00",
-                "Session5": "Race",
-                "Session5Date": "2023-07-02 15:00:00+02:00",
-                "Session5DateUtc": "2023-07-02 13:00:00",
-                "F1ApiSupport": True,
-            },
-            {
-                "RoundNumber": 10,
-                "Country": "Great Britain",
-                "Location": "Silverstone",
-                "OfficialEventName": "FORMULA 1 ARAMCO BRITISH GRAND PRIX 2023",
-                "EventDate": "2023-07-09",
-                "EventName": "British Grand Prix",
-                "EventFormat": "conventional",
-                "Session1": "Practice 1",
-                "Session1Date": "2023-07-07 12:30:00+01:00",
-                "Session1DateUtc": "2023-07-07 11:30:00",
-                "Session2": "Practice 2",
-                "Session2Date": "2023-07-07 16:00:00+01:00",
-                "Session2DateUtc": "2023-07-07 15:00:00",
-                "Session3": "Practice 3",
-                "Session3Date": "2023-07-08 11:30:00+01:00",
-                "Session3DateUtc": "2023-07-08 10:30:00",
-                "Session4": "Qualifying",
-                "Session4Date": "2023-07-08 15:00:00+01:00",
-                "Session4DateUtc": "2023-07-08 14:00:00",
-                "Session5": "Race",
-                "Session5Date": "2023-07-09 15:00:00+01:00",
-                "Session5DateUtc": "2023-07-09 14:00:00",
-                "F1ApiSupport": True,
-            },
-            {
-                "RoundNumber": 11,
-                "Country": "Hungary",
-                "Location": "Budapest",
-                "OfficialEventName": "FORMULA 1 QATAR AIRWAYS HUNGARIAN GRAND PRIX 2023",
-                "EventDate": "2023-07-23",
-                "EventName": "Hungarian Grand Prix",
-                "EventFormat": "conventional",
-                "Session1": "Practice 1",
-                "Session1Date": "2023-07-21 13:30:00+02:00",
-                "Session1DateUtc": "2023-07-21 11:30:00",
-                "Session2": "Practice 2",
-                "Session2Date": "2023-07-21 17:00:00+02:00",
-                "Session2DateUtc": "2023-07-21 15:00:00",
-                "Session3": "Practice 3",
-                "Session3Date": "2023-07-22 12:30:00+02:00",
-                "Session3DateUtc": "2023-07-22 10:30:00",
-                "Session4": "Qualifying",
-                "Session4Date": "2023-07-22 16:00:00+02:00",
-                "Session4DateUtc": "2023-07-22 14:00:00",
-                "Session5": "Race",
-                "Session5Date": "2023-07-23 15:00:00+02:00",
-                "Session5DateUtc": "2023-07-23 13:00:00",
-                "F1ApiSupport": True,
-            },
-            {
-                "RoundNumber": 12,
-                "Country": "Belgium",
-                "Location": "Spa-Francorchamps",
-                "OfficialEventName": "FORMULA 1 MSC CRUISES BELGIAN GRAND PRIX 2023",
-                "EventDate": "2023-07-30",
-                "EventName": "Belgian Grand Prix",
-                "EventFormat": "sprint_shootout",
-                "Session1": "Practice 1",
-                "Session1Date": "2023-07-28 13:30:00+02:00",
-                "Session1DateUtc": "2023-07-28 11:30:00",
-                "Session2": "Qualifying",
-                "Session2Date": "2023-07-28 17:00:00+02:00",
-                "Session2DateUtc": "2023-07-28 15:00:00",
-                "Session3": "Sprint Shootout",
-                "Session3Date": "2023-07-29 12:00:00+02:00",
-                "Session3DateUtc": "2023-07-29 10:00:00",
-                "Session4": "Sprint",
-                "Session4Date": "2023-07-29 17:05:00+02:00",
-                "Session4DateUtc": "2023-07-29 15:05:00",
-                "Session5": "Race",
-                "Session5Date": "2023-07-30 15:00:00+02:00",
-                "Session5DateUtc": "2023-07-30 13:00:00",
-                "F1ApiSupport": True,
-            },
-            {
-                "RoundNumber": 13,
-                "Country": "Netherlands",
-                "Location": "Zandvoort",
-                "OfficialEventName": "FORMULA 1 HEINEKEN DUTCH GRAND PRIX 2023",
-                "EventDate": "2023-08-27",
-                "EventName": "Dutch Grand Prix",
-                "EventFormat": "conventional",
-                "Session1": "Practice 1",
-                "Session1Date": "2023-08-25 12:30:00+02:00",
-                "Session1DateUtc": "2023-08-25 10:30:00",
-                "Session2": "Practice 2",
-                "Session2Date": "2023-08-25 16:00:00+02:00",
-                "Session2DateUtc": "2023-08-25 14:00:00",
-                "Session3": "Practice 3",
-                "Session3Date": "2023-08-26 11:30:00+02:00",
-                "Session3DateUtc": "2023-08-26 09:30:00",
-                "Session4": "Qualifying",
-                "Session4Date": "2023-08-26 15:00:00+02:00",
-                "Session4DateUtc": "2023-08-26 13:00:00",
-                "Session5": "Race",
-                "Session5Date": "2023-08-27 15:00:00+02:00",
-                "Session5DateUtc": "2023-08-27 13:00:00",
-                "F1ApiSupport": True,
-            },
-            {
-                "RoundNumber": 14,
-                "Country": "Italy",
-                "Location": "Monza",
-                "OfficialEventName": "FORMULA 1 PIRELLI GRAN PREMIO D’ITALIA 2023 ",
-                "EventDate": "2023-09-03",
-                "EventName": "Italian Grand Prix",
-                "EventFormat": "conventional",
-                "Session1": "Practice 1",
-                "Session1Date": "2023-09-01 13:30:00+02:00",
-                "Session1DateUtc": "2023-09-01 11:30:00",
-                "Session2": "Practice 2",
-                "Session2Date": "2023-09-01 17:00:00+02:00",
-                "Session2DateUtc": "2023-09-01 15:00:00",
-                "Session3": "Practice 3",
-                "Session3Date": "2023-09-02 12:30:00+02:00",
-                "Session3DateUtc": "2023-09-02 10:30:00",
-                "Session4": "Qualifying",
-                "Session4Date": "2023-09-02 16:00:00+02:00",
-                "Session4DateUtc": "2023-09-02 14:00:00",
-                "Session5": "Race",
-                "Session5Date": "2023-09-03 15:00:00+02:00",
-                "Session5DateUtc": "2023-09-03 13:00:00",
-                "F1ApiSupport": True,
-            },
-            {
-                "RoundNumber": 15,
-                "Country": "Singapore",
-                "Location": "Marina Bay",
-                "OfficialEventName": "FORMULA 1 SINGAPORE AIRLINES SINGAPORE GRAND PRIX 2023 ",
-                "EventDate": "2023-09-17",
-                "EventName": "Singapore Grand Prix",
-                "EventFormat": "conventional",
-                "Session1": "Practice 1",
-                "Session1Date": "2023-09-15 17:30:00+08:00",
-                "Session1DateUtc": "2023-09-15 09:30:00",
-                "Session2": "Practice 2",
-                "Session2Date": "2023-09-15 21:00:00+08:00",
-                "Session2DateUtc": "2023-09-15 13:00:00",
-                "Session3": "Practice 3",
-                "Session3Date": "2023-09-16 17:30:00+08:00",
-                "Session3DateUtc": "2023-09-16 09:30:00",
-                "Session4": "Qualifying",
-                "Session4Date": "2023-09-16 21:00:00+08:00",
-                "Session4DateUtc": "2023-09-16 13:00:00",
-                "Session5": "Race",
-                "Session5Date": "2023-09-17 20:00:00+08:00",
-                "Session5DateUtc": "2023-09-17 12:00:00",
-                "F1ApiSupport": True,
-            },
-            {
-                "RoundNumber": 16,
-                "Country": "Japan",
-                "Location": "Suzuka",
-                "OfficialEventName": "FORMULA 1 LENOVO JAPANESE GRAND PRIX 2023 ",
-                "EventDate": "2023-09-24",
-                "EventName": "Japanese Grand Prix",
-                "EventFormat": "conventional",
-                "Session1": "Practice 1",
-                "Session1Date": "2023-09-22 11:30:00+09:00",
-                "Session1DateUtc": "2023-09-22 02:30:00",
-                "Session2": "Practice 2",
-                "Session2Date": "2023-09-22 15:00:00+09:00",
-                "Session2DateUtc": "2023-09-22 06:00:00",
-                "Session3": "Practice 3",
-                "Session3Date": "2023-09-23 11:30:00+09:00",
-                "Session3DateUtc": "2023-09-23 02:30:00",
-                "Session4": "Qualifying",
-                "Session4Date": "2023-09-23 15:00:00+09:00",
-                "Session4DateUtc": "2023-09-23 06:00:00",
-                "Session5": "Race",
-                "Session5Date": "2023-09-24 14:00:00+09:00",
-                "Session5DateUtc": "2023-09-24 05:00:00",
-                "F1ApiSupport": True,
-            },
-            {
-                "RoundNumber": 17,
-                "Country": "Qatar",
-                "Location": "Lusail",
-                "OfficialEventName": "FORMULA 1 QATAR AIRWAYS QATAR GRAND PRIX 2023",
-                "EventDate": "2023-10-08",
-                "EventName": "Qatar Grand Prix",
-                "EventFormat": "sprint_shootout",
-                "Session1": "Practice 1",
-                "Session1Date": "2023-10-06 16:30:00+03:00",
-                "Session1DateUtc": "2023-10-06 13:30:00",
-                "Session2": "Qualifying",
-                "Session2Date": "2023-10-06 20:00:00+03:00",
-                "Session2DateUtc": "2023-10-06 17:00:00",
-                "Session3": "Sprint Shootout",
-                "Session3Date": "2023-10-07 16:20:00+03:00",
-                "Session3DateUtc": "2023-10-07 13:20:00",
-                "Session4": "Sprint",
-                "Session4Date": "2023-10-07 20:30:00+03:00",
-                "Session4DateUtc": "2023-10-07 17:30:00",
-                "Session5": "Race",
-                "Session5Date": "2023-10-08 20:00:00+03:00",
-                "Session5DateUtc": "2023-10-08 17:00:00",
-                "F1ApiSupport": True,
-            },
-            {
-                "RoundNumber": 18,
-                "Country": "United States",
-                "Location": "Austin",
-                "OfficialEventName": "FORMULA 1 LENOVO UNITED STATES GRAND PRIX 2023",
-                "EventDate": "2023-10-22",
-                "EventName": "United States Grand Prix",
-                "EventFormat": "sprint_shootout",
-                "Session1": "Practice 1",
-                "Session1Date": "2023-10-20 12:30:00-05:00",
-                "Session1DateUtc": "2023-10-20 17:30:00",
-                "Session2": "Qualifying",
-                "Session2Date": "2023-10-20 16:00:00-05:00",
-                "Session2DateUtc": "2023-10-20 21:00:00",
-                "Session3": "Sprint Shootout",
-                "Session3Date": "2023-10-21 12:30:00-05:00",
-                "Session3DateUtc": "2023-10-21 17:30:00",
-                "Session4": "Sprint",
-                "Session4Date": "2023-10-21 17:00:00-05:00",
-                "Session4DateUtc": "2023-10-21 22:00:00",
-                "Session5": "Race",
-                "Session5Date": "2023-10-22 14:00:00-05:00",
-                "Session5DateUtc": "2023-10-22 19:00:00",
-                "F1ApiSupport": True,
-            },
-            {
-                "RoundNumber": 19,
-                "Country": "Mexico",
-                "Location": "Mexico City",
-                "OfficialEventName": "FORMULA 1 GRAN PREMIO DE LA CIUDAD DE MÉXICO 2023",
-                "EventDate": "2023-10-29",
-                "EventName": "Mexico City Grand Prix",
-                "EventFormat": "conventional",
-                "Session1": "Practice 1",
-                "Session1Date": "2023-10-27 12:30:00-06:00",
-                "Session1DateUtc": "2023-10-27 18:30:00",
-                "Session2": "Practice 2",
-                "Session2Date": "2023-10-27 16:00:00-06:00",
-                "Session2DateUtc": "2023-10-27 22:00:00",
-                "Session3": "Practice 3",
-                "Session3Date": "2023-10-28 11:30:00-06:00",
-                "Session3DateUtc": "2023-10-28 17:30:00",
-                "Session4": "Qualifying",
-                "Session4Date": "2023-10-28 15:00:00-06:00",
-                "Session4DateUtc": "2023-10-28 21:00:00",
-                "Session5": "Race",
-                "Session5Date": "2023-10-29 14:00:00-06:00",
-                "Session5DateUtc": "2023-10-29 20:00:00",
-                "F1ApiSupport": True,
-            },
-            {
-                "RoundNumber": 20,
-                "Country": "Brazil",
-                "Location": "São Paulo",
-                "OfficialEventName": "FORMULA 1 ROLEX GRANDE PRÊMIO DE SÃO PAULO 2023",
-                "EventDate": "2023-11-05",
-                "EventName": "São Paulo Grand Prix",
-                "EventFormat": "sprint_shootout",
-                "Session1": "Practice 1",
-                "Session1Date": "2023-11-03 11:30:00-03:00",
-                "Session1DateUtc": "2023-11-03 14:30:00",
-                "Session2": "Qualifying",
-                "Session2Date": "2023-11-03 15:00:00-03:00",
-                "Session2DateUtc": "2023-11-03 18:00:00",
-                "Session3": "Sprint Shootout",
-                "Session3Date": "2023-11-04 11:00:00-03:00",
-                "Session3DateUtc": "2023-11-04 14:00:00",
-                "Session4": "Sprint",
-                "Session4Date": "2023-11-04 15:30:00-03:00",
-                "Session4DateUtc": "2023-11-04 18:30:00",
-                "Session5": "Race",
-                "Session5Date": "2023-11-05 14:00:00-03:00",
-                "Session5DateUtc": "2023-11-05 17:00:00",
-                "F1ApiSupport": True,
-            },
-            {
-                "RoundNumber": 21,
-                "Country": "United States",
-                "Location": "Las Vegas",
-                "OfficialEventName": "FORMULA 1 HEINEKEN SILVER LAS VEGAS GRAND PRIX 2023",
-                "EventDate": "2023-11-18",
-                "EventName": "Las Vegas Grand Prix",
-                "EventFormat": "conventional",
-                "Session1": "Practice 1",
-                "Session1Date": "2023-11-16 20:30:00-08:00",
-                "Session1DateUtc": "2023-11-17 04:30:00",
-                "Session2": "Practice 2",
-                "Session2Date": "2023-11-17 02:30:00-08:00",
-                "Session2DateUtc": "2023-11-17 10:30:00",
-                "Session3": "Practice 3",
-                "Session3Date": "2023-11-17 20:30:00-08:00",
-                "Session3DateUtc": "2023-11-18 04:30:00",
-                "Session4": "Qualifying",
-                "Session4Date": "2023-11-18 00:00:00-08:00",
-                "Session4DateUtc": "2023-11-18 08:00:00",
-                "Session5": "Race",
-                "Session5Date": "2023-11-18 22:00:00-08:00",
-                "Session5DateUtc": "2023-11-19 06:00:00",
-                "F1ApiSupport": True,
-            },
-            {
-                "RoundNumber": 22,
-                "Country": "Abu Dhabi",
-                "Location": "Yas Island",
-                "OfficialEventName": "FORMULA 1 ETIHAD AIRWAYS ABU DHABI GRAND PRIX 2023 ",
-                "EventDate": "2023-11-26",
-                "EventName": "Abu Dhabi Grand Prix",
-                "EventFormat": "conventional",
-                "Session1": "Practice 1",
-                "Session1Date": "2023-11-24 13:30:00+04:00",
-                "Session1DateUtc": "2023-11-24 09:30:00",
-                "Session2": "Practice 2",
-                "Session2Date": "2023-11-24 17:00:00+04:00",
-                "Session2DateUtc": "2023-11-24 13:00:00",
-                "Session3": "Practice 3",
-                "Session3Date": "2023-11-25 14:30:00+04:00",
-                "Session3DateUtc": "2023-11-25 10:30:00",
-                "Session4": "Qualifying",
-                "Session4Date": "2023-11-25 18:00:00+04:00",
-                "Session4DateUtc": "2023-11-25 14:00:00",
-                "Session5": "Race",
-                "Session5Date": "2023-11-26 17:00:00+04:00",
-                "Session5DateUtc": "2023-11-26 13:00:00",
-                "F1ApiSupport": True,
-            },
-        ],
-    }
-
-
-def test_get_schedule_good_year():
-    response = client.get("/schedule?year=2023")
-    assert response.status_code == status.HTTP_200_OK
-    assert response.json() == {
-        "year": 2023,
-        "EventSchedule": [
-            {
-                "RoundNumber": 1,
-                "Country": "Bahrain",
-                "Location": "Sakhir",
-                "OfficialEventName": "FORMULA 1 GULF AIR BAHRAIN GRAND PRIX 2023",
-                "EventDate": "2023-03-05",
-                "EventName": "Bahrain Grand Prix",
-                "EventFormat": "conventional",
-                "Session1": "Practice 1",
-                "Session1Date": "2023-03-03 14:30:00+03:00",
-                "Session1DateUtc": "2023-03-03 11:30:00",
-                "Session2": "Practice 2",
-                "Session2Date": "2023-03-03 18:00:00+03:00",
-                "Session2DateUtc": "2023-03-03 15:00:00",
-                "Session3": "Practice 3",
-                "Session3Date": "2023-03-04 14:30:00+03:00",
-                "Session3DateUtc": "2023-03-04 11:30:00",
-                "Session4": "Qualifying",
-                "Session4Date": "2023-03-04 18:00:00+03:00",
-                "Session4DateUtc": "2023-03-04 15:00:00",
-                "Session5": "Race",
-                "Session5Date": "2023-03-05 18:00:00+03:00",
-                "Session5DateUtc": "2023-03-05 15:00:00",
-                "F1ApiSupport": True,
-            },
-            {
-                "RoundNumber": 2,
-                "Country": "Saudi Arabia",
-                "Location": "Jeddah",
-                "OfficialEventName": "FORMULA 1 STC SAUDI ARABIAN GRAND PRIX 2023",
-                "EventDate": "2023-03-19",
-                "EventName": "Saudi Arabian Grand Prix",
-                "EventFormat": "conventional",
-                "Session1": "Practice 1",
-                "Session1Date": "2023-03-17 16:30:00+03:00",
-                "Session1DateUtc": "2023-03-17 13:30:00",
-                "Session2": "Practice 2",
-                "Session2Date": "2023-03-17 20:00:00+03:00",
-                "Session2DateUtc": "2023-03-17 17:00:00",
-                "Session3": "Practice 3",
-                "Session3Date": "2023-03-18 16:30:00+03:00",
-                "Session3DateUtc": "2023-03-18 13:30:00",
-                "Session4": "Qualifying",
-                "Session4Date": "2023-03-18 20:00:00+03:00",
-                "Session4DateUtc": "2023-03-18 17:00:00",
-                "Session5": "Race",
-                "Session5Date": "2023-03-19 20:00:00+03:00",
-                "Session5DateUtc": "2023-03-19 17:00:00",
-                "F1ApiSupport": True,
-            },
-            {
-                "RoundNumber": 3,
-                "Country": "Australia",
-                "Location": "Melbourne",
-                "OfficialEventName": "FORMULA 1 ROLEX AUSTRALIAN GRAND PRIX 2023",
-                "EventDate": "2023-04-02",
-                "EventName": "Australian Grand Prix",
-                "EventFormat": "conventional",
-                "Session1": "Practice 1",
-                "Session1Date": "2023-03-31 12:30:00+10:00",
-                "Session1DateUtc": "2023-03-31 02:30:00",
-                "Session2": "Practice 2",
-                "Session2Date": "2023-03-31 16:00:00+10:00",
-                "Session2DateUtc": "2023-03-31 06:00:00",
-                "Session3": "Practice 3",
-                "Session3Date": "2023-04-01 12:30:00+10:00",
-                "Session3DateUtc": "2023-04-01 02:30:00",
-                "Session4": "Qualifying",
-                "Session4Date": "2023-04-01 16:00:00+10:00",
-                "Session4DateUtc": "2023-04-01 06:00:00",
-                "Session5": "Race",
-                "Session5Date": "2023-04-02 15:00:00+10:00",
-                "Session5DateUtc": "2023-04-02 05:00:00",
-                "F1ApiSupport": True,
-            },
-            {
-                "RoundNumber": 4,
-                "Country": "Azerbaijan",
-                "Location": "Baku",
-                "OfficialEventName": "FORMULA 1 AZERBAIJAN GRAND PRIX 2023",
-                "EventDate": "2023-04-30",
-                "EventName": "Azerbaijan Grand Prix",
-                "EventFormat": "sprint_shootout",
-                "Session1": "Practice 1",
-                "Session1Date": "2023-04-28 13:30:00+04:00",
-                "Session1DateUtc": "2023-04-28 09:30:00",
-                "Session2": "Qualifying",
-                "Session2Date": "2023-04-28 17:00:00+04:00",
-                "Session2DateUtc": "2023-04-28 13:00:00",
-                "Session3": "Sprint Shootout",
-                "Session3Date": "2023-04-29 12:30:00+04:00",
-                "Session3DateUtc": "2023-04-29 08:30:00",
-                "Session4": "Sprint",
-                "Session4Date": "2023-04-29 17:30:00+04:00",
-                "Session4DateUtc": "2023-04-29 13:30:00",
-                "Session5": "Race",
-                "Session5Date": "2023-04-30 15:00:00+04:00",
-                "Session5DateUtc": "2023-04-30 11:00:00",
-                "F1ApiSupport": True,
-            },
-            {
-                "RoundNumber": 5,
-                "Country": "United States",
-                "Location": "Miami",
-                "OfficialEventName": "FORMULA 1 CRYPTO.COM MIAMI GRAND PRIX 2023",
-                "EventDate": "2023-05-07",
-                "EventName": "Miami Grand Prix",
-                "EventFormat": "conventional",
-                "Session1": "Practice 1",
-                "Session1Date": "2023-05-05 14:00:00-04:00",
-                "Session1DateUtc": "2023-05-05 18:00:00",
-                "Session2": "Practice 2",
-                "Session2Date": "2023-05-05 17:30:00-04:00",
-                "Session2DateUtc": "2023-05-05 21:30:00",
-                "Session3": "Practice 3",
-                "Session3Date": "2023-05-06 12:30:00-04:00",
-                "Session3DateUtc": "2023-05-06 16:30:00",
-                "Session4": "Qualifying",
-                "Session4Date": "2023-05-06 16:00:00-04:00",
-                "Session4DateUtc": "2023-05-06 20:00:00",
-                "Session5": "Race",
-                "Session5Date": "2023-05-07 15:30:00-04:00",
-                "Session5DateUtc": "2023-05-07 19:30:00",
-                "F1ApiSupport": True,
-            },
-            {
-                "RoundNumber": 6,
-                "Country": "Monaco",
-                "Location": "Monaco",
-                "OfficialEventName": "FORMULA 1 GRAND PRIX DE MONACO 2023",
-                "EventDate": "2023-05-28",
-                "EventName": "Monaco Grand Prix",
-                "EventFormat": "conventional",
-                "Session1": "Practice 1",
-                "Session1Date": "2023-05-26 13:30:00+02:00",
-                "Session1DateUtc": "2023-05-26 11:30:00",
-                "Session2": "Practice 2",
-                "Session2Date": "2023-05-26 17:00:00+02:00",
-                "Session2DateUtc": "2023-05-26 15:00:00",
-                "Session3": "Practice 3",
-                "Session3Date": "2023-05-27 12:30:00+02:00",
-                "Session3DateUtc": "2023-05-27 10:30:00",
-                "Session4": "Qualifying",
-                "Session4Date": "2023-05-27 16:00:00+02:00",
-                "Session4DateUtc": "2023-05-27 14:00:00",
-                "Session5": "Race",
-                "Session5Date": "2023-05-28 15:00:00+02:00",
-                "Session5DateUtc": "2023-05-28 13:00:00",
-                "F1ApiSupport": True,
-            },
-            {
-                "RoundNumber": 7,
-                "Country": "Spain",
-                "Location": "Barcelona",
-                "OfficialEventName": "FORMULA 1 AWS GRAN PREMIO DE ESPAÑA 2023",
-                "EventDate": "2023-06-04",
-                "EventName": "Spanish Grand Prix",
-                "EventFormat": "conventional",
-                "Session1": "Practice 1",
-                "Session1Date": "2023-06-02 13:30:00+02:00",
-                "Session1DateUtc": "2023-06-02 11:30:00",
-                "Session2": "Practice 2",
-                "Session2Date": "2023-06-02 17:00:00+02:00",
-                "Session2DateUtc": "2023-06-02 15:00:00",
-                "Session3": "Practice 3",
-                "Session3Date": "2023-06-03 12:30:00+02:00",
-                "Session3DateUtc": "2023-06-03 10:30:00",
-                "Session4": "Qualifying",
-                "Session4Date": "2023-06-03 16:00:00+02:00",
-                "Session4DateUtc": "2023-06-03 14:00:00",
-                "Session5": "Race",
-                "Session5Date": "2023-06-04 15:00:00+02:00",
-                "Session5DateUtc": "2023-06-04 13:00:00",
-                "F1ApiSupport": True,
-            },
-            {
-                "RoundNumber": 8,
-                "Country": "Canada",
-                "Location": "Montréal",
-                "OfficialEventName": "FORMULA 1 PIRELLI GRAND PRIX DU CANADA 2023",
-                "EventDate": "2023-06-18",
-                "EventName": "Canadian Grand Prix",
-                "EventFormat": "conventional",
-                "Session1": "Practice 1",
-                "Session1Date": "2023-06-16 13:30:00-04:00",
-                "Session1DateUtc": "2023-06-16 17:30:00",
-                "Session2": "Practice 2",
-                "Session2Date": "2023-06-16 16:30:00-04:00",
-                "Session2DateUtc": "2023-06-16 20:30:00",
-                "Session3": "Practice 3",
-                "Session3Date": "2023-06-17 12:30:00-04:00",
-                "Session3DateUtc": "2023-06-17 16:30:00",
-                "Session4": "Qualifying",
-                "Session4Date": "2023-06-17 16:00:00-04:00",
-                "Session4DateUtc": "2023-06-17 20:00:00",
-                "Session5": "Race",
-                "Session5Date": "2023-06-18 14:00:00-04:00",
-                "Session5DateUtc": "2023-06-18 18:00:00",
-                "F1ApiSupport": True,
-            },
-            {
-                "RoundNumber": 9,
-                "Country": "Austria",
-                "Location": "Spielberg",
-                "OfficialEventName": "FORMULA 1 ROLEX GROSSER PREIS VON ÖSTERREICH 2023",
-                "EventDate": "2023-07-02",
-                "EventName": "Austrian Grand Prix",
-                "EventFormat": "sprint_shootout",
-                "Session1": "Practice 1",
-                "Session1Date": "2023-06-30 13:30:00+02:00",
-                "Session1DateUtc": "2023-06-30 11:30:00",
-                "Session2": "Qualifying",
-                "Session2Date": "2023-06-30 17:00:00+02:00",
-                "Session2DateUtc": "2023-06-30 15:00:00",
-                "Session3": "Sprint Shootout",
-                "Session3Date": "2023-07-01 12:00:00+02:00",
-                "Session3DateUtc": "2023-07-01 10:00:00",
-                "Session4": "Sprint",
-                "Session4Date": "2023-07-01 16:30:00+02:00",
-                "Session4DateUtc": "2023-07-01 14:30:00",
-                "Session5": "Race",
-                "Session5Date": "2023-07-02 15:00:00+02:00",
-                "Session5DateUtc": "2023-07-02 13:00:00",
-                "F1ApiSupport": True,
-            },
-            {
-                "RoundNumber": 10,
-                "Country": "Great Britain",
-                "Location": "Silverstone",
-                "OfficialEventName": "FORMULA 1 ARAMCO BRITISH GRAND PRIX 2023",
-                "EventDate": "2023-07-09",
-                "EventName": "British Grand Prix",
-                "EventFormat": "conventional",
-                "Session1": "Practice 1",
-                "Session1Date": "2023-07-07 12:30:00+01:00",
-                "Session1DateUtc": "2023-07-07 11:30:00",
-                "Session2": "Practice 2",
-                "Session2Date": "2023-07-07 16:00:00+01:00",
-                "Session2DateUtc": "2023-07-07 15:00:00",
-                "Session3": "Practice 3",
-                "Session3Date": "2023-07-08 11:30:00+01:00",
-                "Session3DateUtc": "2023-07-08 10:30:00",
-                "Session4": "Qualifying",
-                "Session4Date": "2023-07-08 15:00:00+01:00",
-                "Session4DateUtc": "2023-07-08 14:00:00",
-                "Session5": "Race",
-                "Session5Date": "2023-07-09 15:00:00+01:00",
-                "Session5DateUtc": "2023-07-09 14:00:00",
-                "F1ApiSupport": True,
-            },
-            {
-                "RoundNumber": 11,
-                "Country": "Hungary",
-                "Location": "Budapest",
-                "OfficialEventName": "FORMULA 1 QATAR AIRWAYS HUNGARIAN GRAND PRIX 2023",
-                "EventDate": "2023-07-23",
-                "EventName": "Hungarian Grand Prix",
-                "EventFormat": "conventional",
-                "Session1": "Practice 1",
-                "Session1Date": "2023-07-21 13:30:00+02:00",
-                "Session1DateUtc": "2023-07-21 11:30:00",
-                "Session2": "Practice 2",
-                "Session2Date": "2023-07-21 17:00:00+02:00",
-                "Session2DateUtc": "2023-07-21 15:00:00",
-                "Session3": "Practice 3",
-                "Session3Date": "2023-07-22 12:30:00+02:00",
-                "Session3DateUtc": "2023-07-22 10:30:00",
-                "Session4": "Qualifying",
-                "Session4Date": "2023-07-22 16:00:00+02:00",
-                "Session4DateUtc": "2023-07-22 14:00:00",
-                "Session5": "Race",
-                "Session5Date": "2023-07-23 15:00:00+02:00",
-                "Session5DateUtc": "2023-07-23 13:00:00",
-                "F1ApiSupport": True,
-            },
-            {
-                "RoundNumber": 12,
-                "Country": "Belgium",
-                "Location": "Spa-Francorchamps",
-                "OfficialEventName": "FORMULA 1 MSC CRUISES BELGIAN GRAND PRIX 2023",
-                "EventDate": "2023-07-30",
-                "EventName": "Belgian Grand Prix",
-                "EventFormat": "sprint_shootout",
-                "Session1": "Practice 1",
-                "Session1Date": "2023-07-28 13:30:00+02:00",
-                "Session1DateUtc": "2023-07-28 11:30:00",
-                "Session2": "Qualifying",
-                "Session2Date": "2023-07-28 17:00:00+02:00",
-                "Session2DateUtc": "2023-07-28 15:00:00",
-                "Session3": "Sprint Shootout",
-                "Session3Date": "2023-07-29 12:00:00+02:00",
-                "Session3DateUtc": "2023-07-29 10:00:00",
-                "Session4": "Sprint",
-                "Session4Date": "2023-07-29 17:05:00+02:00",
-                "Session4DateUtc": "2023-07-29 15:05:00",
-                "Session5": "Race",
-                "Session5Date": "2023-07-30 15:00:00+02:00",
-                "Session5DateUtc": "2023-07-30 13:00:00",
-                "F1ApiSupport": True,
-            },
-            {
-                "RoundNumber": 13,
-                "Country": "Netherlands",
-                "Location": "Zandvoort",
-                "OfficialEventName": "FORMULA 1 HEINEKEN DUTCH GRAND PRIX 2023",
-                "EventDate": "2023-08-27",
-                "EventName": "Dutch Grand Prix",
-                "EventFormat": "conventional",
-                "Session1": "Practice 1",
-                "Session1Date": "2023-08-25 12:30:00+02:00",
-                "Session1DateUtc": "2023-08-25 10:30:00",
-                "Session2": "Practice 2",
-                "Session2Date": "2023-08-25 16:00:00+02:00",
-                "Session2DateUtc": "2023-08-25 14:00:00",
-                "Session3": "Practice 3",
-                "Session3Date": "2023-08-26 11:30:00+02:00",
-                "Session3DateUtc": "2023-08-26 09:30:00",
-                "Session4": "Qualifying",
-                "Session4Date": "2023-08-26 15:00:00+02:00",
-                "Session4DateUtc": "2023-08-26 13:00:00",
-                "Session5": "Race",
-                "Session5Date": "2023-08-27 15:00:00+02:00",
-                "Session5DateUtc": "2023-08-27 13:00:00",
-                "F1ApiSupport": True,
-            },
-            {
-                "RoundNumber": 14,
-                "Country": "Italy",
-                "Location": "Monza",
-                "OfficialEventName": "FORMULA 1 PIRELLI GRAN PREMIO D’ITALIA 2023 ",
-                "EventDate": "2023-09-03",
-                "EventName": "Italian Grand Prix",
-                "EventFormat": "conventional",
-                "Session1": "Practice 1",
-                "Session1Date": "2023-09-01 13:30:00+02:00",
-                "Session1DateUtc": "2023-09-01 11:30:00",
-                "Session2": "Practice 2",
-                "Session2Date": "2023-09-01 17:00:00+02:00",
-                "Session2DateUtc": "2023-09-01 15:00:00",
-                "Session3": "Practice 3",
-                "Session3Date": "2023-09-02 12:30:00+02:00",
-                "Session3DateUtc": "2023-09-02 10:30:00",
-                "Session4": "Qualifying",
-                "Session4Date": "2023-09-02 16:00:00+02:00",
-                "Session4DateUtc": "2023-09-02 14:00:00",
-                "Session5": "Race",
-                "Session5Date": "2023-09-03 15:00:00+02:00",
-                "Session5DateUtc": "2023-09-03 13:00:00",
-                "F1ApiSupport": True,
-            },
-            {
-                "RoundNumber": 15,
-                "Country": "Singapore",
-                "Location": "Marina Bay",
-                "OfficialEventName": "FORMULA 1 SINGAPORE AIRLINES SINGAPORE GRAND PRIX 2023 ",
-                "EventDate": "2023-09-17",
-                "EventName": "Singapore Grand Prix",
-                "EventFormat": "conventional",
-                "Session1": "Practice 1",
-                "Session1Date": "2023-09-15 17:30:00+08:00",
-                "Session1DateUtc": "2023-09-15 09:30:00",
-                "Session2": "Practice 2",
-                "Session2Date": "2023-09-15 21:00:00+08:00",
-                "Session2DateUtc": "2023-09-15 13:00:00",
-                "Session3": "Practice 3",
-                "Session3Date": "2023-09-16 17:30:00+08:00",
-                "Session3DateUtc": "2023-09-16 09:30:00",
-                "Session4": "Qualifying",
-                "Session4Date": "2023-09-16 21:00:00+08:00",
-                "Session4DateUtc": "2023-09-16 13:00:00",
-                "Session5": "Race",
-                "Session5Date": "2023-09-17 20:00:00+08:00",
-                "Session5DateUtc": "2023-09-17 12:00:00",
-                "F1ApiSupport": True,
-            },
-            {
-                "RoundNumber": 16,
-                "Country": "Japan",
-                "Location": "Suzuka",
-                "OfficialEventName": "FORMULA 1 LENOVO JAPANESE GRAND PRIX 2023 ",
-                "EventDate": "2023-09-24",
-                "EventName": "Japanese Grand Prix",
-                "EventFormat": "conventional",
-                "Session1": "Practice 1",
-                "Session1Date": "2023-09-22 11:30:00+09:00",
-                "Session1DateUtc": "2023-09-22 02:30:00",
-                "Session2": "Practice 2",
-                "Session2Date": "2023-09-22 15:00:00+09:00",
-                "Session2DateUtc": "2023-09-22 06:00:00",
-                "Session3": "Practice 3",
-                "Session3Date": "2023-09-23 11:30:00+09:00",
-                "Session3DateUtc": "2023-09-23 02:30:00",
-                "Session4": "Qualifying",
-                "Session4Date": "2023-09-23 15:00:00+09:00",
-                "Session4DateUtc": "2023-09-23 06:00:00",
-                "Session5": "Race",
-                "Session5Date": "2023-09-24 14:00:00+09:00",
-                "Session5DateUtc": "2023-09-24 05:00:00",
-                "F1ApiSupport": True,
-            },
-            {
-                "RoundNumber": 17,
-                "Country": "Qatar",
-                "Location": "Lusail",
-                "OfficialEventName": "FORMULA 1 QATAR AIRWAYS QATAR GRAND PRIX 2023",
-                "EventDate": "2023-10-08",
-                "EventName": "Qatar Grand Prix",
-                "EventFormat": "sprint_shootout",
-                "Session1": "Practice 1",
-                "Session1Date": "2023-10-06 16:30:00+03:00",
-                "Session1DateUtc": "2023-10-06 13:30:00",
-                "Session2": "Qualifying",
-                "Session2Date": "2023-10-06 20:00:00+03:00",
-                "Session2DateUtc": "2023-10-06 17:00:00",
-                "Session3": "Sprint Shootout",
-                "Session3Date": "2023-10-07 16:20:00+03:00",
-                "Session3DateUtc": "2023-10-07 13:20:00",
-                "Session4": "Sprint",
-                "Session4Date": "2023-10-07 20:30:00+03:00",
-                "Session4DateUtc": "2023-10-07 17:30:00",
-                "Session5": "Race",
-                "Session5Date": "2023-10-08 20:00:00+03:00",
-                "Session5DateUtc": "2023-10-08 17:00:00",
-                "F1ApiSupport": True,
-            },
-            {
-                "RoundNumber": 18,
-                "Country": "United States",
-                "Location": "Austin",
-                "OfficialEventName": "FORMULA 1 LENOVO UNITED STATES GRAND PRIX 2023",
-                "EventDate": "2023-10-22",
-                "EventName": "United States Grand Prix",
-                "EventFormat": "sprint_shootout",
-                "Session1": "Practice 1",
-                "Session1Date": "2023-10-20 12:30:00-05:00",
-                "Session1DateUtc": "2023-10-20 17:30:00",
-                "Session2": "Qualifying",
-                "Session2Date": "2023-10-20 16:00:00-05:00",
-                "Session2DateUtc": "2023-10-20 21:00:00",
-                "Session3": "Sprint Shootout",
-                "Session3Date": "2023-10-21 12:30:00-05:00",
-                "Session3DateUtc": "2023-10-21 17:30:00",
-                "Session4": "Sprint",
-                "Session4Date": "2023-10-21 17:00:00-05:00",
-                "Session4DateUtc": "2023-10-21 22:00:00",
-                "Session5": "Race",
-                "Session5Date": "2023-10-22 14:00:00-05:00",
-                "Session5DateUtc": "2023-10-22 19:00:00",
-                "F1ApiSupport": True,
-            },
-            {
-                "RoundNumber": 19,
-                "Country": "Mexico",
-                "Location": "Mexico City",
-                "OfficialEventName": "FORMULA 1 GRAN PREMIO DE LA CIUDAD DE MÉXICO 2023",
-                "EventDate": "2023-10-29",
-                "EventName": "Mexico City Grand Prix",
-                "EventFormat": "conventional",
-                "Session1": "Practice 1",
-                "Session1Date": "2023-10-27 12:30:00-06:00",
-                "Session1DateUtc": "2023-10-27 18:30:00",
-                "Session2": "Practice 2",
-                "Session2Date": "2023-10-27 16:00:00-06:00",
-                "Session2DateUtc": "2023-10-27 22:00:00",
-                "Session3": "Practice 3",
-                "Session3Date": "2023-10-28 11:30:00-06:00",
-                "Session3DateUtc": "2023-10-28 17:30:00",
-                "Session4": "Qualifying",
-                "Session4Date": "2023-10-28 15:00:00-06:00",
-                "Session4DateUtc": "2023-10-28 21:00:00",
-                "Session5": "Race",
-                "Session5Date": "2023-10-29 14:00:00-06:00",
-                "Session5DateUtc": "2023-10-29 20:00:00",
-                "F1ApiSupport": True,
-            },
-            {
-                "RoundNumber": 20,
-                "Country": "Brazil",
-                "Location": "São Paulo",
-                "OfficialEventName": "FORMULA 1 ROLEX GRANDE PRÊMIO DE SÃO PAULO 2023",
-                "EventDate": "2023-11-05",
-                "EventName": "São Paulo Grand Prix",
-                "EventFormat": "sprint_shootout",
-                "Session1": "Practice 1",
-                "Session1Date": "2023-11-03 11:30:00-03:00",
-                "Session1DateUtc": "2023-11-03 14:30:00",
-                "Session2": "Qualifying",
-                "Session2Date": "2023-11-03 15:00:00-03:00",
-                "Session2DateUtc": "2023-11-03 18:00:00",
-                "Session3": "Sprint Shootout",
-                "Session3Date": "2023-11-04 11:00:00-03:00",
-                "Session3DateUtc": "2023-11-04 14:00:00",
-                "Session4": "Sprint",
-                "Session4Date": "2023-11-04 15:30:00-03:00",
-                "Session4DateUtc": "2023-11-04 18:30:00",
-                "Session5": "Race",
-                "Session5Date": "2023-11-05 14:00:00-03:00",
-                "Session5DateUtc": "2023-11-05 17:00:00",
-                "F1ApiSupport": True,
-            },
-            {
-                "RoundNumber": 21,
-                "Country": "United States",
-                "Location": "Las Vegas",
-                "OfficialEventName": "FORMULA 1 HEINEKEN SILVER LAS VEGAS GRAND PRIX 2023",
-                "EventDate": "2023-11-18",
-                "EventName": "Las Vegas Grand Prix",
-                "EventFormat": "conventional",
-                "Session1": "Practice 1",
-                "Session1Date": "2023-11-16 20:30:00-08:00",
-                "Session1DateUtc": "2023-11-17 04:30:00",
-                "Session2": "Practice 2",
-                "Session2Date": "2023-11-17 02:30:00-08:00",
-                "Session2DateUtc": "2023-11-17 10:30:00",
-                "Session3": "Practice 3",
-                "Session3Date": "2023-11-17 20:30:00-08:00",
-                "Session3DateUtc": "2023-11-18 04:30:00",
-                "Session4": "Qualifying",
-                "Session4Date": "2023-11-18 00:00:00-08:00",
-                "Session4DateUtc": "2023-11-18 08:00:00",
-                "Session5": "Race",
-                "Session5Date": "2023-11-18 22:00:00-08:00",
-                "Session5DateUtc": "2023-11-19 06:00:00",
-                "F1ApiSupport": True,
-            },
-            {
-                "RoundNumber": 22,
-                "Country": "Abu Dhabi",
-                "Location": "Yas Island",
-                "OfficialEventName": "FORMULA 1 ETIHAD AIRWAYS ABU DHABI GRAND PRIX 2023 ",
-                "EventDate": "2023-11-26",
-                "EventName": "Abu Dhabi Grand Prix",
-                "EventFormat": "conventional",
-                "Session1": "Practice 1",
-                "Session1Date": "2023-11-24 13:30:00+04:00",
-                "Session1DateUtc": "2023-11-24 09:30:00",
-                "Session2": "Practice 2",
-                "Session2Date": "2023-11-24 17:00:00+04:00",
-                "Session2DateUtc": "2023-11-24 13:00:00",
-                "Session3": "Practice 3",
-                "Session3Date": "2023-11-25 14:30:00+04:00",
-                "Session3DateUtc": "2023-11-25 10:30:00",
-                "Session4": "Qualifying",
-                "Session4Date": "2023-11-25 18:00:00+04:00",
-                "Session4DateUtc": "2023-11-25 14:00:00",
-                "Session5": "Race",
-                "Session5Date": "2023-11-26 17:00:00+04:00",
-                "Session5DateUtc": "2023-11-26 13:00:00",
-                "F1ApiSupport": True,
-            },
-        ],
-    }
-
-
-# endregion schedule - good inputs
-
-# endregion schedule
-
-
-# region next-event
-
-
-def test_get_next_event():
-    response = client.get("/next-event")
-    assert response.status_code == status.HTTP_200_OK
-    assert response.json() == {
-        "RoundNumber": 1,
-        "Country": "Bahrain",
-        "Location": "Sakhir",
-        "OfficialEventName": "FORMULA 1 GULF AIR BAHRAIN GRAND PRIX 2024",
-        "EventDate": "2024-03-02 00:00:00",
-        "EventName": "Bahrain Grand Prix",
-        "EventFormat": "conventional",
-        "Session1": "Practice 1",
-        "Session1Date": "2024-02-29 14:30:00+03:00",
-        "Session1DateUtc": "2024-02-29 11:30:00",
-        "Session2": "Practice 2",
-        "Session2Date": "2024-02-29 18:00:00+03:00",
-        "Session2DateUtc": "2024-02-29 15:00:00",
-        "Session3": "Practice 3",
-        "Session3Date": "2024-03-01 15:30:00+03:00",
-        "Session3DateUtc": "2024-03-01 12:30:00",
-        "Session4": "Qualifying",
-        "Session4Date": "2024-03-01 19:00:00+03:00",
-        "Session4DateUtc": "2024-03-01 16:00:00",
-        "Session5": "Race",
-        "Session5Date": "2024-03-02 18:00:00+03:00",
-        "Session5DateUtc": "2024-03-02 15:00:00",
-        "F1ApiSupport": True,
-    }
-
-
-# endregion next-event
-
-
-# region standings
-
-# region standings - good inputs
-
-
-def test_get_standings():
-    response = client.get("/standings")
-    assert response.status_code == status.HTTP_200_OK
-    assert response.json() == {
-        "season": 2023,
-        "round": 22,
-        "DriverStandings": [
-            {
-                "position": "1",
-                "positionText": "1",
-                "points": "575",
-                "wins": "19",
-                "Driver": {
-                    "driverId": "max_verstappen",
-                    "permanentNumber": "33",
-                    "code": "VER",
-                    "url": "http://en.wikipedia.org/wiki/Max_Verstappen",
-                    "givenName": "Max",
-                    "familyName": "Verstappen",
-                    "dateOfBirth": "1997-09-30",
-                    "nationality": "Dutch",
-                },
-                "Constructors": [
-                    {
-                        "constructorId": "red_bull",
-                        "url": "http://en.wikipedia.org/wiki/Red_Bull_Racing",
-                        "name": "Red Bull",
-                        "nationality": "Austrian",
-                    }
-                ],
-            },
-            {
-                "position": "2",
-                "positionText": "2",
-                "points": "285",
-                "wins": "2",
-                "Driver": {
-                    "driverId": "perez",
-                    "permanentNumber": "11",
-                    "code": "PER",
-                    "url": "http://en.wikipedia.org/wiki/Sergio_P%C3%A9rez",
-                    "givenName": "Sergio",
-                    "familyName": "Pérez",
-                    "dateOfBirth": "1990-01-26",
-                    "nationality": "Mexican",
-                },
-                "Constructors": [
-                    {
-                        "constructorId": "red_bull",
-                        "url": "http://en.wikipedia.org/wiki/Red_Bull_Racing",
-                        "name": "Red Bull",
-                        "nationality": "Austrian",
-                    }
-                ],
-            },
-            {
-                "position": "3",
-                "positionText": "3",
-                "points": "234",
-                "wins": "0",
-                "Driver": {
-                    "driverId": "hamilton",
-                    "permanentNumber": "44",
-                    "code": "HAM",
-                    "url": "http://en.wikipedia.org/wiki/Lewis_Hamilton",
-                    "givenName": "Lewis",
-                    "familyName": "Hamilton",
-                    "dateOfBirth": "1985-01-07",
-                    "nationality": "British",
-                },
-                "Constructors": [
-                    {
-                        "constructorId": "mercedes",
-                        "url": "http://en.wikipedia.org/wiki/Mercedes-Benz_in_Formula_One",
-                        "name": "Mercedes",
-                        "nationality": "German",
-                    }
-                ],
-            },
-            {
-                "position": "4",
-                "positionText": "4",
-                "points": "206",
-                "wins": "0",
-                "Driver": {
-                    "driverId": "alonso",
-                    "permanentNumber": "14",
-                    "code": "ALO",
-                    "url": "http://en.wikipedia.org/wiki/Fernando_Alonso",
-                    "givenName": "Fernando",
-                    "familyName": "Alonso",
-                    "dateOfBirth": "1981-07-29",
-                    "nationality": "Spanish",
-                },
-                "Constructors": [
-                    {
-                        "constructorId": "aston_martin",
-                        "url": "http://en.wikipedia.org/wiki/Aston_Martin_in_Formula_One",
-                        "name": "Aston Martin",
-                        "nationality": "British",
-                    }
-                ],
-            },
-            {
-                "position": "5",
-                "positionText": "5",
-                "points": "206",
-                "wins": "0",
-                "Driver": {
-                    "driverId": "leclerc",
-                    "permanentNumber": "16",
-                    "code": "LEC",
-                    "url": "http://en.wikipedia.org/wiki/Charles_Leclerc",
-                    "givenName": "Charles",
-                    "familyName": "Leclerc",
-                    "dateOfBirth": "1997-10-16",
-                    "nationality": "Monegasque",
-                },
-                "Constructors": [
-                    {
-                        "constructorId": "ferrari",
-                        "url": "http://en.wikipedia.org/wiki/Scuderia_Ferrari",
-                        "name": "Ferrari",
-                        "nationality": "Italian",
-                    }
-                ],
-            },
-            {
-                "position": "6",
-                "positionText": "6",
-                "points": "205",
-                "wins": "0",
-                "Driver": {
-                    "driverId": "norris",
-                    "permanentNumber": "4",
-                    "code": "NOR",
-                    "url": "http://en.wikipedia.org/wiki/Lando_Norris",
-                    "givenName": "Lando",
-                    "familyName": "Norris",
-                    "dateOfBirth": "1999-11-13",
-                    "nationality": "British",
-                },
-                "Constructors": [
-                    {
-                        "constructorId": "mclaren",
-                        "url": "http://en.wikipedia.org/wiki/McLaren",
-                        "name": "McLaren",
-                        "nationality": "British",
-                    }
-                ],
-            },
-            {
-                "position": "7",
-                "positionText": "7",
-                "points": "200",
-                "wins": "1",
-                "Driver": {
-                    "driverId": "sainz",
-                    "permanentNumber": "55",
-                    "code": "SAI",
-                    "url": "http://en.wikipedia.org/wiki/Carlos_Sainz_Jr.",
-                    "givenName": "Carlos",
-                    "familyName": "Sainz",
-                    "dateOfBirth": "1994-09-01",
-                    "nationality": "Spanish",
-                },
-                "Constructors": [
-                    {
-                        "constructorId": "ferrari",
-                        "url": "http://en.wikipedia.org/wiki/Scuderia_Ferrari",
-                        "name": "Ferrari",
-                        "nationality": "Italian",
-                    }
-                ],
-            },
-            {
-                "position": "8",
-                "positionText": "8",
-                "points": "175",
-                "wins": "0",
-                "Driver": {
-                    "driverId": "russell",
-                    "permanentNumber": "63",
-                    "code": "RUS",
-                    "url": "http://en.wikipedia.org/wiki/George_Russell_(racing_driver)",
-                    "givenName": "George",
-                    "familyName": "Russell",
-                    "dateOfBirth": "1998-02-15",
-                    "nationality": "British",
-                },
-                "Constructors": [
-                    {
-                        "constructorId": "mercedes",
-                        "url": "http://en.wikipedia.org/wiki/Mercedes-Benz_in_Formula_One",
-                        "name": "Mercedes",
-                        "nationality": "German",
-                    }
-                ],
-            },
-            {
-                "position": "9",
-                "positionText": "9",
-                "points": "97",
-                "wins": "0",
-                "Driver": {
-                    "driverId": "piastri",
-                    "permanentNumber": "81",
-                    "code": "PIA",
-                    "url": "http://en.wikipedia.org/wiki/Oscar_Piastri",
-                    "givenName": "Oscar",
-                    "familyName": "Piastri",
-                    "dateOfBirth": "2001-04-06",
-                    "nationality": "Australian",
-                },
-                "Constructors": [
-                    {
-                        "constructorId": "mclaren",
-                        "url": "http://en.wikipedia.org/wiki/McLaren",
-                        "name": "McLaren",
-                        "nationality": "British",
-                    }
-                ],
-            },
-            {
-                "position": "10",
-                "positionText": "10",
-                "points": "74",
-                "wins": "0",
-                "Driver": {
-                    "driverId": "stroll",
-                    "permanentNumber": "18",
-                    "code": "STR",
-                    "url": "http://en.wikipedia.org/wiki/Lance_Stroll",
-                    "givenName": "Lance",
-                    "familyName": "Stroll",
-                    "dateOfBirth": "1998-10-29",
-                    "nationality": "Canadian",
-                },
-                "Constructors": [
-                    {
-                        "constructorId": "aston_martin",
-                        "url": "http://en.wikipedia.org/wiki/Aston_Martin_in_Formula_One",
-                        "name": "Aston Martin",
-                        "nationality": "British",
-                    }
-                ],
-            },
-            {
-                "position": "11",
-                "positionText": "11",
-                "points": "62",
-                "wins": "0",
-                "Driver": {
-                    "driverId": "gasly",
-                    "permanentNumber": "10",
-                    "code": "GAS",
-                    "url": "http://en.wikipedia.org/wiki/Pierre_Gasly",
-                    "givenName": "Pierre",
-                    "familyName": "Gasly",
-                    "dateOfBirth": "1996-02-07",
-                    "nationality": "French",
-                },
-                "Constructors": [
-                    {
-                        "constructorId": "alpine",
-                        "url": "http://en.wikipedia.org/wiki/Alpine_F1_Team",
-                        "name": "Alpine F1 Team",
-                        "nationality": "French",
-                    }
-                ],
-            },
-            {
-                "position": "12",
-                "positionText": "12",
-                "points": "58",
-                "wins": "0",
-                "Driver": {
-                    "driverId": "ocon",
-                    "permanentNumber": "31",
-                    "code": "OCO",
-                    "url": "http://en.wikipedia.org/wiki/Esteban_Ocon",
-                    "givenName": "Esteban",
-                    "familyName": "Ocon",
-                    "dateOfBirth": "1996-09-17",
-                    "nationality": "French",
-                },
-                "Constructors": [
-                    {
-                        "constructorId": "alpine",
-                        "url": "http://en.wikipedia.org/wiki/Alpine_F1_Team",
-                        "name": "Alpine F1 Team",
-                        "nationality": "French",
-                    }
-                ],
-            },
-            {
-                "position": "13",
-                "positionText": "13",
-                "points": "27",
-                "wins": "0",
-                "Driver": {
-                    "driverId": "albon",
-                    "permanentNumber": "23",
-                    "code": "ALB",
-                    "url": "http://en.wikipedia.org/wiki/Alexander_Albon",
-                    "givenName": "Alexander",
-                    "familyName": "Albon",
-                    "dateOfBirth": "1996-03-23",
-                    "nationality": "Thai",
-                },
-                "Constructors": [
-                    {
-                        "constructorId": "williams",
-                        "url": "http://en.wikipedia.org/wiki/Williams_Grand_Prix_Engineering",
-                        "name": "Williams",
-                        "nationality": "British",
-                    }
-                ],
-            },
-            {
-                "position": "14",
-                "positionText": "14",
-                "points": "17",
-                "wins": "0",
-                "Driver": {
-                    "driverId": "tsunoda",
-                    "permanentNumber": "22",
-                    "code": "TSU",
-                    "url": "http://en.wikipedia.org/wiki/Yuki_Tsunoda",
-                    "givenName": "Yuki",
-                    "familyName": "Tsunoda",
-                    "dateOfBirth": "2000-05-11",
-                    "nationality": "Japanese",
-                },
-                "Constructors": [
-                    {
-                        "constructorId": "alphatauri",
-                        "url": "http://en.wikipedia.org/wiki/Scuderia_AlphaTauri",
-                        "name": "AlphaTauri",
-                        "nationality": "Italian",
-                    }
-                ],
-            },
-            {
-                "position": "15",
-                "positionText": "15",
-                "points": "10",
-                "wins": "0",
-                "Driver": {
-                    "driverId": "bottas",
-                    "permanentNumber": "77",
-                    "code": "BOT",
-                    "url": "http://en.wikipedia.org/wiki/Valtteri_Bottas",
-                    "givenName": "Valtteri",
-                    "familyName": "Bottas",
-                    "dateOfBirth": "1989-08-28",
-                    "nationality": "Finnish",
-                },
-                "Constructors": [
-                    {
-                        "constructorId": "alfa",
-                        "url": "http://en.wikipedia.org/wiki/Alfa_Romeo_in_Formula_One",
-                        "name": "Alfa Romeo",
-                        "nationality": "Swiss",
-                    }
-                ],
-            },
-            {
-                "position": "16",
-                "positionText": "16",
-                "points": "9",
-                "wins": "0",
-                "Driver": {
-                    "driverId": "hulkenberg",
-                    "permanentNumber": "27",
-                    "code": "HUL",
-                    "url": "http://en.wikipedia.org/wiki/Nico_H%C3%BClkenberg",
-                    "givenName": "Nico",
-                    "familyName": "Hülkenberg",
-                    "dateOfBirth": "1987-08-19",
-                    "nationality": "German",
-                },
-                "Constructors": [
-                    {
-                        "constructorId": "haas",
-                        "url": "http://en.wikipedia.org/wiki/Haas_F1_Team",
-                        "name": "Haas F1 Team",
-                        "nationality": "American",
-                    }
-                ],
-            },
-            {
-                "position": "17",
-                "positionText": "17",
-                "points": "6",
-                "wins": "0",
-                "Driver": {
-                    "driverId": "ricciardo",
-                    "permanentNumber": "3",
-                    "code": "RIC",
-                    "url": "http://en.wikipedia.org/wiki/Daniel_Ricciardo",
-                    "givenName": "Daniel",
-                    "familyName": "Ricciardo",
-                    "dateOfBirth": "1989-07-01",
-                    "nationality": "Australian",
-                },
-                "Constructors": [
-                    {
-                        "constructorId": "alphatauri",
-                        "url": "http://en.wikipedia.org/wiki/Scuderia_AlphaTauri",
-                        "name": "AlphaTauri",
-                        "nationality": "Italian",
-                    }
-                ],
-            },
-            {
-                "position": "18",
-                "positionText": "18",
-                "points": "6",
-                "wins": "0",
-                "Driver": {
-                    "driverId": "zhou",
-                    "permanentNumber": "24",
-                    "code": "ZHO",
-                    "url": "http://en.wikipedia.org/wiki/Zhou_Guanyu",
-                    "givenName": "Guanyu",
-                    "familyName": "Zhou",
-                    "dateOfBirth": "1999-05-30",
-                    "nationality": "Chinese",
-                },
-                "Constructors": [
-                    {
-                        "constructorId": "alfa",
-                        "url": "http://en.wikipedia.org/wiki/Alfa_Romeo_in_Formula_One",
-                        "name": "Alfa Romeo",
-                        "nationality": "Swiss",
-                    }
-                ],
-            },
-            {
-                "position": "19",
-                "positionText": "19",
-                "points": "3",
-                "wins": "0",
-                "Driver": {
-                    "driverId": "kevin_magnussen",
-                    "permanentNumber": "20",
-                    "code": "MAG",
-                    "url": "http://en.wikipedia.org/wiki/Kevin_Magnussen",
-                    "givenName": "Kevin",
-                    "familyName": "Magnussen",
-                    "dateOfBirth": "1992-10-05",
-                    "nationality": "Danish",
-                },
-                "Constructors": [
-                    {
-                        "constructorId": "haas",
-                        "url": "http://en.wikipedia.org/wiki/Haas_F1_Team",
-                        "name": "Haas F1 Team",
-                        "nationality": "American",
-                    }
-                ],
-            },
-            {
-                "position": "20",
-                "positionText": "20",
-                "points": "2",
-                "wins": "0",
-                "Driver": {
-                    "driverId": "lawson",
-                    "permanentNumber": "40",
-                    "code": "LAW",
-                    "url": "http://en.wikipedia.org/wiki/Liam_Lawson",
-                    "givenName": "Liam",
-                    "familyName": "Lawson",
-                    "dateOfBirth": "2002-02-11",
-                    "nationality": "New Zealander",
-                },
-                "Constructors": [
-                    {
-                        "constructorId": "alphatauri",
-                        "url": "http://en.wikipedia.org/wiki/Scuderia_AlphaTauri",
-                        "name": "AlphaTauri",
-                        "nationality": "Italian",
-                    }
-                ],
-            },
-            {
-                "position": "21",
-                "positionText": "21",
-                "points": "1",
-                "wins": "0",
-                "Driver": {
-                    "driverId": "sargeant",
-                    "permanentNumber": "2",
-                    "code": "SAR",
-                    "url": "http://en.wikipedia.org/wiki/Logan_Sargeant",
-                    "givenName": "Logan",
-                    "familyName": "Sargeant",
-                    "dateOfBirth": "2000-12-31",
-                    "nationality": "American",
-                },
-                "Constructors": [
-                    {
-                        "constructorId": "williams",
-                        "url": "http://en.wikipedia.org/wiki/Williams_Grand_Prix_Engineering",
-                        "name": "Williams",
-                        "nationality": "British",
-                    }
-                ],
-            },
-            {
-                "position": "22",
-                "positionText": "22",
-                "points": "0",
-                "wins": "0",
-                "Driver": {
-                    "driverId": "de_vries",
-                    "permanentNumber": "21",
-                    "code": "DEV",
-                    "url": "http://en.wikipedia.org/wiki/Nyck_de_Vries",
-                    "givenName": "Nyck",
-                    "familyName": "de Vries",
-                    "dateOfBirth": "1995-02-06",
-                    "nationality": "Dutch",
-                },
-                "Constructors": [
-                    {
-                        "constructorId": "alphatauri",
-                        "url": "http://en.wikipedia.org/wiki/Scuderia_AlphaTauri",
-                        "name": "AlphaTauri",
-                        "nationality": "Italian",
-                    }
-                ],
-            },
-        ],
-        "ConstructorStandings": [
-            {
-                "position": "1",
-                "positionText": "1",
-                "points": "860",
-                "wins": "21",
-                "Constructor": {
-                    "constructorId": "red_bull",
-                    "url": "http://en.wikipedia.org/wiki/Red_Bull_Racing",
-                    "name": "Red Bull",
-                    "nationality": "Austrian",
-                },
-            },
-            {
-                "position": "2",
-                "positionText": "2",
-                "points": "409",
-                "wins": "0",
-                "Constructor": {
-                    "constructorId": "mercedes",
-                    "url": "http://en.wikipedia.org/wiki/Mercedes-Benz_in_Formula_One",
-                    "name": "Mercedes",
-                    "nationality": "German",
-                },
-            },
-            {
-                "position": "3",
-                "positionText": "3",
-                "points": "406",
-                "wins": "1",
-                "Constructor": {
-                    "constructorId": "ferrari",
-                    "url": "http://en.wikipedia.org/wiki/Scuderia_Ferrari",
-                    "name": "Ferrari",
-                    "nationality": "Italian",
-                },
-            },
-            {
-                "position": "4",
-                "positionText": "4",
-                "points": "302",
-                "wins": "0",
-                "Constructor": {
-                    "constructorId": "mclaren",
-                    "url": "http://en.wikipedia.org/wiki/McLaren",
-                    "name": "McLaren",
-                    "nationality": "British",
-                },
-            },
-            {
-                "position": "5",
-                "positionText": "5",
-                "points": "280",
-                "wins": "0",
-                "Constructor": {
-                    "constructorId": "aston_martin",
-                    "url": "http://en.wikipedia.org/wiki/Aston_Martin_in_Formula_One",
-                    "name": "Aston Martin",
-                    "nationality": "British",
-                },
-            },
-            {
-                "position": "6",
-                "positionText": "6",
-                "points": "120",
-                "wins": "0",
-                "Constructor": {
-                    "constructorId": "alpine",
-                    "url": "http://en.wikipedia.org/wiki/Alpine_F1_Team",
-                    "name": "Alpine F1 Team",
-                    "nationality": "French",
-                },
-            },
-            {
-                "position": "7",
-                "positionText": "7",
-                "points": "28",
-                "wins": "0",
-                "Constructor": {
-                    "constructorId": "williams",
-                    "url": "http://en.wikipedia.org/wiki/Williams_Grand_Prix_Engineering",
-                    "name": "Williams",
-                    "nationality": "British",
-                },
-            },
-            {
-                "position": "8",
-                "positionText": "8",
-                "points": "25",
-                "wins": "0",
-                "Constructor": {
-                    "constructorId": "alphatauri",
-                    "url": "http://en.wikipedia.org/wiki/Scuderia_AlphaTauri",
-                    "name": "AlphaTauri",
-                    "nationality": "Italian",
-                },
-            },
-            {
-                "position": "9",
-                "positionText": "9",
-                "points": "16",
-                "wins": "0",
-                "Constructor": {
-                    "constructorId": "alfa",
-                    "url": "http://en.wikipedia.org/wiki/Alfa_Romeo_in_Formula_One",
-                    "name": "Alfa Romeo",
-                    "nationality": "Swiss",
-                },
-            },
-            {
-                "position": "10",
-                "positionText": "10",
-                "points": "12",
-                "wins": "0",
-                "Constructor": {
-                    "constructorId": "haas",
-                    "url": "http://en.wikipedia.org/wiki/Haas_F1_Team",
-                    "name": "Haas F1 Team",
-                    "nationality": "American",
-                },
-            },
-        ],
-    }
-
-
-def test_get_standings_good_year_only():
-    response = client.get("/standings?year=2023")
-    assert response.status_code == status.HTTP_200_OK
-    assert response.json() == {
-        "season": 2023,
-        "round": 22,
-        "DriverStandings": [
-            {
-                "position": "1",
-                "positionText": "1",
-                "points": "575",
-                "wins": "19",
-                "Driver": {
-                    "driverId": "max_verstappen",
-                    "permanentNumber": "33",
-                    "code": "VER",
-                    "url": "http://en.wikipedia.org/wiki/Max_Verstappen",
-                    "givenName": "Max",
-                    "familyName": "Verstappen",
-                    "dateOfBirth": "1997-09-30",
-                    "nationality": "Dutch",
-                },
-                "Constructors": [
-                    {
-                        "constructorId": "red_bull",
-                        "url": "http://en.wikipedia.org/wiki/Red_Bull_Racing",
-                        "name": "Red Bull",
-                        "nationality": "Austrian",
-                    }
-                ],
-            },
-            {
-                "position": "2",
-                "positionText": "2",
-                "points": "285",
-                "wins": "2",
-                "Driver": {
-                    "driverId": "perez",
-                    "permanentNumber": "11",
-                    "code": "PER",
-                    "url": "http://en.wikipedia.org/wiki/Sergio_P%C3%A9rez",
-                    "givenName": "Sergio",
-                    "familyName": "Pérez",
-                    "dateOfBirth": "1990-01-26",
-                    "nationality": "Mexican",
-                },
-                "Constructors": [
-                    {
-                        "constructorId": "red_bull",
-                        "url": "http://en.wikipedia.org/wiki/Red_Bull_Racing",
-                        "name": "Red Bull",
-                        "nationality": "Austrian",
-                    }
-                ],
-            },
-            {
-                "position": "3",
-                "positionText": "3",
-                "points": "234",
-                "wins": "0",
-                "Driver": {
-                    "driverId": "hamilton",
-                    "permanentNumber": "44",
-                    "code": "HAM",
-                    "url": "http://en.wikipedia.org/wiki/Lewis_Hamilton",
-                    "givenName": "Lewis",
-                    "familyName": "Hamilton",
-                    "dateOfBirth": "1985-01-07",
-                    "nationality": "British",
-                },
-                "Constructors": [
-                    {
-                        "constructorId": "mercedes",
-                        "url": "http://en.wikipedia.org/wiki/Mercedes-Benz_in_Formula_One",
-                        "name": "Mercedes",
-                        "nationality": "German",
-                    }
-                ],
-            },
-            {
-                "position": "4",
-                "positionText": "4",
-                "points": "206",
-                "wins": "0",
-                "Driver": {
-                    "driverId": "alonso",
-                    "permanentNumber": "14",
-                    "code": "ALO",
-                    "url": "http://en.wikipedia.org/wiki/Fernando_Alonso",
-                    "givenName": "Fernando",
-                    "familyName": "Alonso",
-                    "dateOfBirth": "1981-07-29",
-                    "nationality": "Spanish",
-                },
-                "Constructors": [
-                    {
-                        "constructorId": "aston_martin",
-                        "url": "http://en.wikipedia.org/wiki/Aston_Martin_in_Formula_One",
-                        "name": "Aston Martin",
-                        "nationality": "British",
-                    }
-                ],
-            },
-            {
-                "position": "5",
-                "positionText": "5",
-                "points": "206",
-                "wins": "0",
-                "Driver": {
-                    "driverId": "leclerc",
-                    "permanentNumber": "16",
-                    "code": "LEC",
-                    "url": "http://en.wikipedia.org/wiki/Charles_Leclerc",
-                    "givenName": "Charles",
-                    "familyName": "Leclerc",
-                    "dateOfBirth": "1997-10-16",
-                    "nationality": "Monegasque",
-                },
-                "Constructors": [
-                    {
-                        "constructorId": "ferrari",
-                        "url": "http://en.wikipedia.org/wiki/Scuderia_Ferrari",
-                        "name": "Ferrari",
-                        "nationality": "Italian",
-                    }
-                ],
-            },
-            {
-                "position": "6",
-                "positionText": "6",
-                "points": "205",
-                "wins": "0",
-                "Driver": {
-                    "driverId": "norris",
-                    "permanentNumber": "4",
-                    "code": "NOR",
-                    "url": "http://en.wikipedia.org/wiki/Lando_Norris",
-                    "givenName": "Lando",
-                    "familyName": "Norris",
-                    "dateOfBirth": "1999-11-13",
-                    "nationality": "British",
-                },
-                "Constructors": [
-                    {
-                        "constructorId": "mclaren",
-                        "url": "http://en.wikipedia.org/wiki/McLaren",
-                        "name": "McLaren",
-                        "nationality": "British",
-                    }
-                ],
-            },
-            {
-                "position": "7",
-                "positionText": "7",
-                "points": "200",
-                "wins": "1",
-                "Driver": {
-                    "driverId": "sainz",
-                    "permanentNumber": "55",
-                    "code": "SAI",
-                    "url": "http://en.wikipedia.org/wiki/Carlos_Sainz_Jr.",
-                    "givenName": "Carlos",
-                    "familyName": "Sainz",
-                    "dateOfBirth": "1994-09-01",
-                    "nationality": "Spanish",
-                },
-                "Constructors": [
-                    {
-                        "constructorId": "ferrari",
-                        "url": "http://en.wikipedia.org/wiki/Scuderia_Ferrari",
-                        "name": "Ferrari",
-                        "nationality": "Italian",
-                    }
-                ],
-            },
-            {
-                "position": "8",
-                "positionText": "8",
-                "points": "175",
-                "wins": "0",
-                "Driver": {
-                    "driverId": "russell",
-                    "permanentNumber": "63",
-                    "code": "RUS",
-                    "url": "http://en.wikipedia.org/wiki/George_Russell_(racing_driver)",
-                    "givenName": "George",
-                    "familyName": "Russell",
-                    "dateOfBirth": "1998-02-15",
-                    "nationality": "British",
-                },
-                "Constructors": [
-                    {
-                        "constructorId": "mercedes",
-                        "url": "http://en.wikipedia.org/wiki/Mercedes-Benz_in_Formula_One",
-                        "name": "Mercedes",
-                        "nationality": "German",
-                    }
-                ],
-            },
-            {
-                "position": "9",
-                "positionText": "9",
-                "points": "97",
-                "wins": "0",
-                "Driver": {
-                    "driverId": "piastri",
-                    "permanentNumber": "81",
-                    "code": "PIA",
-                    "url": "http://en.wikipedia.org/wiki/Oscar_Piastri",
-                    "givenName": "Oscar",
-                    "familyName": "Piastri",
-                    "dateOfBirth": "2001-04-06",
-                    "nationality": "Australian",
-                },
-                "Constructors": [
-                    {
-                        "constructorId": "mclaren",
-                        "url": "http://en.wikipedia.org/wiki/McLaren",
-                        "name": "McLaren",
-                        "nationality": "British",
-                    }
-                ],
-            },
-            {
-                "position": "10",
-                "positionText": "10",
-                "points": "74",
-                "wins": "0",
-                "Driver": {
-                    "driverId": "stroll",
-                    "permanentNumber": "18",
-                    "code": "STR",
-                    "url": "http://en.wikipedia.org/wiki/Lance_Stroll",
-                    "givenName": "Lance",
-                    "familyName": "Stroll",
-                    "dateOfBirth": "1998-10-29",
-                    "nationality": "Canadian",
-                },
-                "Constructors": [
-                    {
-                        "constructorId": "aston_martin",
-                        "url": "http://en.wikipedia.org/wiki/Aston_Martin_in_Formula_One",
-                        "name": "Aston Martin",
-                        "nationality": "British",
-                    }
-                ],
-            },
-            {
-                "position": "11",
-                "positionText": "11",
-                "points": "62",
-                "wins": "0",
-                "Driver": {
-                    "driverId": "gasly",
-                    "permanentNumber": "10",
-                    "code": "GAS",
-                    "url": "http://en.wikipedia.org/wiki/Pierre_Gasly",
-                    "givenName": "Pierre",
-                    "familyName": "Gasly",
-                    "dateOfBirth": "1996-02-07",
-                    "nationality": "French",
-                },
-                "Constructors": [
-                    {
-                        "constructorId": "alpine",
-                        "url": "http://en.wikipedia.org/wiki/Alpine_F1_Team",
-                        "name": "Alpine F1 Team",
-                        "nationality": "French",
-                    }
-                ],
-            },
-            {
-                "position": "12",
-                "positionText": "12",
-                "points": "58",
-                "wins": "0",
-                "Driver": {
-                    "driverId": "ocon",
-                    "permanentNumber": "31",
-                    "code": "OCO",
-                    "url": "http://en.wikipedia.org/wiki/Esteban_Ocon",
-                    "givenName": "Esteban",
-                    "familyName": "Ocon",
-                    "dateOfBirth": "1996-09-17",
-                    "nationality": "French",
-                },
-                "Constructors": [
-                    {
-                        "constructorId": "alpine",
-                        "url": "http://en.wikipedia.org/wiki/Alpine_F1_Team",
-                        "name": "Alpine F1 Team",
-                        "nationality": "French",
-                    }
-                ],
-            },
-            {
-                "position": "13",
-                "positionText": "13",
-                "points": "27",
-                "wins": "0",
-                "Driver": {
-                    "driverId": "albon",
-                    "permanentNumber": "23",
-                    "code": "ALB",
-                    "url": "http://en.wikipedia.org/wiki/Alexander_Albon",
-                    "givenName": "Alexander",
-                    "familyName": "Albon",
-                    "dateOfBirth": "1996-03-23",
-                    "nationality": "Thai",
-                },
-                "Constructors": [
-                    {
-                        "constructorId": "williams",
-                        "url": "http://en.wikipedia.org/wiki/Williams_Grand_Prix_Engineering",
-                        "name": "Williams",
-                        "nationality": "British",
-                    }
-                ],
-            },
-            {
-                "position": "14",
-                "positionText": "14",
-                "points": "17",
-                "wins": "0",
-                "Driver": {
-                    "driverId": "tsunoda",
-                    "permanentNumber": "22",
-                    "code": "TSU",
-                    "url": "http://en.wikipedia.org/wiki/Yuki_Tsunoda",
-                    "givenName": "Yuki",
-                    "familyName": "Tsunoda",
-                    "dateOfBirth": "2000-05-11",
-                    "nationality": "Japanese",
-                },
-                "Constructors": [
-                    {
-                        "constructorId": "alphatauri",
-                        "url": "http://en.wikipedia.org/wiki/Scuderia_AlphaTauri",
-                        "name": "AlphaTauri",
-                        "nationality": "Italian",
-                    }
-                ],
-            },
-            {
-                "position": "15",
-                "positionText": "15",
-                "points": "10",
-                "wins": "0",
-                "Driver": {
-                    "driverId": "bottas",
-                    "permanentNumber": "77",
-                    "code": "BOT",
-                    "url": "http://en.wikipedia.org/wiki/Valtteri_Bottas",
-                    "givenName": "Valtteri",
-                    "familyName": "Bottas",
-                    "dateOfBirth": "1989-08-28",
-                    "nationality": "Finnish",
-                },
-                "Constructors": [
-                    {
-                        "constructorId": "alfa",
-                        "url": "http://en.wikipedia.org/wiki/Alfa_Romeo_in_Formula_One",
-                        "name": "Alfa Romeo",
-                        "nationality": "Swiss",
-                    }
-                ],
-            },
-            {
-                "position": "16",
-                "positionText": "16",
-                "points": "9",
-                "wins": "0",
-                "Driver": {
-                    "driverId": "hulkenberg",
-                    "permanentNumber": "27",
-                    "code": "HUL",
-                    "url": "http://en.wikipedia.org/wiki/Nico_H%C3%BClkenberg",
-                    "givenName": "Nico",
-                    "familyName": "Hülkenberg",
-                    "dateOfBirth": "1987-08-19",
-                    "nationality": "German",
-                },
-                "Constructors": [
-                    {
-                        "constructorId": "haas",
-                        "url": "http://en.wikipedia.org/wiki/Haas_F1_Team",
-                        "name": "Haas F1 Team",
-                        "nationality": "American",
-                    }
-                ],
-            },
-            {
-                "position": "17",
-                "positionText": "17",
-                "points": "6",
-                "wins": "0",
-                "Driver": {
-                    "driverId": "ricciardo",
-                    "permanentNumber": "3",
-                    "code": "RIC",
-                    "url": "http://en.wikipedia.org/wiki/Daniel_Ricciardo",
-                    "givenName": "Daniel",
-                    "familyName": "Ricciardo",
-                    "dateOfBirth": "1989-07-01",
-                    "nationality": "Australian",
-                },
-                "Constructors": [
-                    {
-                        "constructorId": "alphatauri",
-                        "url": "http://en.wikipedia.org/wiki/Scuderia_AlphaTauri",
-                        "name": "AlphaTauri",
-                        "nationality": "Italian",
-                    }
-                ],
-            },
-            {
-                "position": "18",
-                "positionText": "18",
-                "points": "6",
-                "wins": "0",
-                "Driver": {
-                    "driverId": "zhou",
-                    "permanentNumber": "24",
-                    "code": "ZHO",
-                    "url": "http://en.wikipedia.org/wiki/Zhou_Guanyu",
-                    "givenName": "Guanyu",
-                    "familyName": "Zhou",
-                    "dateOfBirth": "1999-05-30",
-                    "nationality": "Chinese",
-                },
-                "Constructors": [
-                    {
-                        "constructorId": "alfa",
-                        "url": "http://en.wikipedia.org/wiki/Alfa_Romeo_in_Formula_One",
-                        "name": "Alfa Romeo",
-                        "nationality": "Swiss",
-                    }
-                ],
-            },
-            {
-                "position": "19",
-                "positionText": "19",
-                "points": "3",
-                "wins": "0",
-                "Driver": {
-                    "driverId": "kevin_magnussen",
-                    "permanentNumber": "20",
-                    "code": "MAG",
-                    "url": "http://en.wikipedia.org/wiki/Kevin_Magnussen",
-                    "givenName": "Kevin",
-                    "familyName": "Magnussen",
-                    "dateOfBirth": "1992-10-05",
-                    "nationality": "Danish",
-                },
-                "Constructors": [
-                    {
-                        "constructorId": "haas",
-                        "url": "http://en.wikipedia.org/wiki/Haas_F1_Team",
-                        "name": "Haas F1 Team",
-                        "nationality": "American",
-                    }
-                ],
-            },
-            {
-                "position": "20",
-                "positionText": "20",
-                "points": "2",
-                "wins": "0",
-                "Driver": {
-                    "driverId": "lawson",
-                    "permanentNumber": "40",
-                    "code": "LAW",
-                    "url": "http://en.wikipedia.org/wiki/Liam_Lawson",
-                    "givenName": "Liam",
-                    "familyName": "Lawson",
-                    "dateOfBirth": "2002-02-11",
-                    "nationality": "New Zealander",
-                },
-                "Constructors": [
-                    {
-                        "constructorId": "alphatauri",
-                        "url": "http://en.wikipedia.org/wiki/Scuderia_AlphaTauri",
-                        "name": "AlphaTauri",
-                        "nationality": "Italian",
-                    }
-                ],
-            },
-            {
-                "position": "21",
-                "positionText": "21",
-                "points": "1",
-                "wins": "0",
-                "Driver": {
-                    "driverId": "sargeant",
-                    "permanentNumber": "2",
-                    "code": "SAR",
-                    "url": "http://en.wikipedia.org/wiki/Logan_Sargeant",
-                    "givenName": "Logan",
-                    "familyName": "Sargeant",
-                    "dateOfBirth": "2000-12-31",
-                    "nationality": "American",
-                },
-                "Constructors": [
-                    {
-                        "constructorId": "williams",
-                        "url": "http://en.wikipedia.org/wiki/Williams_Grand_Prix_Engineering",
-                        "name": "Williams",
-                        "nationality": "British",
-                    }
-                ],
-            },
-            {
-                "position": "22",
-                "positionText": "22",
-                "points": "0",
-                "wins": "0",
-                "Driver": {
-                    "driverId": "de_vries",
-                    "permanentNumber": "21",
-                    "code": "DEV",
-                    "url": "http://en.wikipedia.org/wiki/Nyck_de_Vries",
-                    "givenName": "Nyck",
-                    "familyName": "de Vries",
-                    "dateOfBirth": "1995-02-06",
-                    "nationality": "Dutch",
-                },
-                "Constructors": [
-                    {
-                        "constructorId": "alphatauri",
-                        "url": "http://en.wikipedia.org/wiki/Scuderia_AlphaTauri",
-                        "name": "AlphaTauri",
-                        "nationality": "Italian",
-                    }
-                ],
-            },
-        ],
-        "ConstructorStandings": [
-            {
-                "position": "1",
-                "positionText": "1",
-                "points": "860",
-                "wins": "21",
-                "Constructor": {
-                    "constructorId": "red_bull",
-                    "url": "http://en.wikipedia.org/wiki/Red_Bull_Racing",
-                    "name": "Red Bull",
-                    "nationality": "Austrian",
-                },
-            },
-            {
-                "position": "2",
-                "positionText": "2",
-                "points": "409",
-                "wins": "0",
-                "Constructor": {
-                    "constructorId": "mercedes",
-                    "url": "http://en.wikipedia.org/wiki/Mercedes-Benz_in_Formula_One",
-                    "name": "Mercedes",
-                    "nationality": "German",
-                },
-            },
-            {
-                "position": "3",
-                "positionText": "3",
-                "points": "406",
-                "wins": "1",
-                "Constructor": {
-                    "constructorId": "ferrari",
-                    "url": "http://en.wikipedia.org/wiki/Scuderia_Ferrari",
-                    "name": "Ferrari",
-                    "nationality": "Italian",
-                },
-            },
-            {
-                "position": "4",
-                "positionText": "4",
-                "points": "302",
-                "wins": "0",
-                "Constructor": {
-                    "constructorId": "mclaren",
-                    "url": "http://en.wikipedia.org/wiki/McLaren",
-                    "name": "McLaren",
-                    "nationality": "British",
-                },
-            },
-            {
-                "position": "5",
-                "positionText": "5",
-                "points": "280",
-                "wins": "0",
-                "Constructor": {
-                    "constructorId": "aston_martin",
-                    "url": "http://en.wikipedia.org/wiki/Aston_Martin_in_Formula_One",
-                    "name": "Aston Martin",
-                    "nationality": "British",
-                },
-            },
-            {
-                "position": "6",
-                "positionText": "6",
-                "points": "120",
-                "wins": "0",
-                "Constructor": {
-                    "constructorId": "alpine",
-                    "url": "http://en.wikipedia.org/wiki/Alpine_F1_Team",
-                    "name": "Alpine F1 Team",
-                    "nationality": "French",
-                },
-            },
-            {
-                "position": "7",
-                "positionText": "7",
-                "points": "28",
-                "wins": "0",
-                "Constructor": {
-                    "constructorId": "williams",
-                    "url": "http://en.wikipedia.org/wiki/Williams_Grand_Prix_Engineering",
-                    "name": "Williams",
-                    "nationality": "British",
-                },
-            },
-            {
-                "position": "8",
-                "positionText": "8",
-                "points": "25",
-                "wins": "0",
-                "Constructor": {
-                    "constructorId": "alphatauri",
-                    "url": "http://en.wikipedia.org/wiki/Scuderia_AlphaTauri",
-                    "name": "AlphaTauri",
-                    "nationality": "Italian",
-                },
-            },
-            {
-                "position": "9",
-                "positionText": "9",
-                "points": "16",
-                "wins": "0",
-                "Constructor": {
-                    "constructorId": "alfa",
-                    "url": "http://en.wikipedia.org/wiki/Alfa_Romeo_in_Formula_One",
-                    "name": "Alfa Romeo",
-                    "nationality": "Swiss",
-                },
-            },
-            {
-                "position": "10",
-                "positionText": "10",
-                "points": "12",
-                "wins": "0",
-                "Constructor": {
-                    "constructorId": "haas",
-                    "url": "http://en.wikipedia.org/wiki/Haas_F1_Team",
-                    "name": "Haas F1 Team",
-                    "nationality": "American",
-                },
-            },
-        ],
-    }
-
-
-def test_get_standings_good_year_and_round():
-    response = client.get("/standings?year=2023&round=5")
-    assert response.status_code == status.HTTP_200_OK
-    assert response.json() == {
-        "season": 2023,
-        "round": 5,
-        "DriverStandings": [
-            {
-                "position": "1",
-                "positionText": "1",
-                "points": "119",
-                "wins": "3",
-                "Driver": {
-                    "driverId": "max_verstappen",
-                    "permanentNumber": "33",
-                    "code": "VER",
-                    "url": "http://en.wikipedia.org/wiki/Max_Verstappen",
-                    "givenName": "Max",
-                    "familyName": "Verstappen",
-                    "dateOfBirth": "1997-09-30",
-                    "nationality": "Dutch",
-                },
-                "Constructors": [
-                    {
-                        "constructorId": "red_bull",
-                        "url": "http://en.wikipedia.org/wiki/Red_Bull_Racing",
-                        "name": "Red Bull",
-                        "nationality": "Austrian",
-                    }
-                ],
-            },
-            {
-                "position": "2",
-                "positionText": "2",
-                "points": "105",
-                "wins": "2",
-                "Driver": {
-                    "driverId": "perez",
-                    "permanentNumber": "11",
-                    "code": "PER",
-                    "url": "http://en.wikipedia.org/wiki/Sergio_P%C3%A9rez",
-                    "givenName": "Sergio",
-                    "familyName": "Pérez",
-                    "dateOfBirth": "1990-01-26",
-                    "nationality": "Mexican",
-                },
-                "Constructors": [
-                    {
-                        "constructorId": "red_bull",
-                        "url": "http://en.wikipedia.org/wiki/Red_Bull_Racing",
-                        "name": "Red Bull",
-                        "nationality": "Austrian",
-                    }
-                ],
-            },
-            {
-                "position": "3",
-                "positionText": "3",
-                "points": "75",
-                "wins": "0",
-                "Driver": {
-                    "driverId": "alonso",
-                    "permanentNumber": "14",
-                    "code": "ALO",
-                    "url": "http://en.wikipedia.org/wiki/Fernando_Alonso",
-                    "givenName": "Fernando",
-                    "familyName": "Alonso",
-                    "dateOfBirth": "1981-07-29",
-                    "nationality": "Spanish",
-                },
-                "Constructors": [
-                    {
-                        "constructorId": "aston_martin",
-                        "url": "http://en.wikipedia.org/wiki/Aston_Martin_in_Formula_One",
-                        "name": "Aston Martin",
-                        "nationality": "British",
-                    }
-                ],
-            },
-            {
-                "position": "4",
-                "positionText": "4",
-                "points": "56",
-                "wins": "0",
-                "Driver": {
-                    "driverId": "hamilton",
-                    "permanentNumber": "44",
-                    "code": "HAM",
-                    "url": "http://en.wikipedia.org/wiki/Lewis_Hamilton",
-                    "givenName": "Lewis",
-                    "familyName": "Hamilton",
-                    "dateOfBirth": "1985-01-07",
-                    "nationality": "British",
-                },
-                "Constructors": [
-                    {
-                        "constructorId": "mercedes",
-                        "url": "http://en.wikipedia.org/wiki/Mercedes-Benz_in_Formula_One",
-                        "name": "Mercedes",
-                        "nationality": "German",
-                    }
-                ],
-            },
-            {
-                "position": "5",
-                "positionText": "5",
-                "points": "44",
-                "wins": "0",
-                "Driver": {
-                    "driverId": "sainz",
-                    "permanentNumber": "55",
-                    "code": "SAI",
-                    "url": "http://en.wikipedia.org/wiki/Carlos_Sainz_Jr.",
-                    "givenName": "Carlos",
-                    "familyName": "Sainz",
-                    "dateOfBirth": "1994-09-01",
-                    "nationality": "Spanish",
-                },
-                "Constructors": [
-                    {
-                        "constructorId": "ferrari",
-                        "url": "http://en.wikipedia.org/wiki/Scuderia_Ferrari",
-                        "name": "Ferrari",
-                        "nationality": "Italian",
-                    }
-                ],
-            },
-            {
-                "position": "6",
-                "positionText": "6",
-                "points": "40",
-                "wins": "0",
-                "Driver": {
-                    "driverId": "russell",
-                    "permanentNumber": "63",
-                    "code": "RUS",
-                    "url": "http://en.wikipedia.org/wiki/George_Russell_(racing_driver)",
-                    "givenName": "George",
-                    "familyName": "Russell",
-                    "dateOfBirth": "1998-02-15",
-                    "nationality": "British",
-                },
-                "Constructors": [
-                    {
-                        "constructorId": "mercedes",
-                        "url": "http://en.wikipedia.org/wiki/Mercedes-Benz_in_Formula_One",
-                        "name": "Mercedes",
-                        "nationality": "German",
-                    }
-                ],
-            },
-            {
-                "position": "7",
-                "positionText": "7",
-                "points": "34",
-                "wins": "0",
-                "Driver": {
-                    "driverId": "leclerc",
-                    "permanentNumber": "16",
-                    "code": "LEC",
-                    "url": "http://en.wikipedia.org/wiki/Charles_Leclerc",
-                    "givenName": "Charles",
-                    "familyName": "Leclerc",
-                    "dateOfBirth": "1997-10-16",
-                    "nationality": "Monegasque",
-                },
-                "Constructors": [
-                    {
-                        "constructorId": "ferrari",
-                        "url": "http://en.wikipedia.org/wiki/Scuderia_Ferrari",
-                        "name": "Ferrari",
-                        "nationality": "Italian",
-                    }
-                ],
-            },
-            {
-                "position": "8",
-                "positionText": "8",
-                "points": "27",
-                "wins": "0",
-                "Driver": {
-                    "driverId": "stroll",
-                    "permanentNumber": "18",
-                    "code": "STR",
-                    "url": "http://en.wikipedia.org/wiki/Lance_Stroll",
-                    "givenName": "Lance",
-                    "familyName": "Stroll",
-                    "dateOfBirth": "1998-10-29",
-                    "nationality": "Canadian",
-                },
-                "Constructors": [
-                    {
-                        "constructorId": "aston_martin",
-                        "url": "http://en.wikipedia.org/wiki/Aston_Martin_in_Formula_One",
-                        "name": "Aston Martin",
-                        "nationality": "British",
-                    }
-                ],
-            },
-            {
-                "position": "9",
-                "positionText": "9",
-                "points": "10",
-                "wins": "0",
-                "Driver": {
-                    "driverId": "norris",
-                    "permanentNumber": "4",
-                    "code": "NOR",
-                    "url": "http://en.wikipedia.org/wiki/Lando_Norris",
-                    "givenName": "Lando",
-                    "familyName": "Norris",
-                    "dateOfBirth": "1999-11-13",
-                    "nationality": "British",
-                },
-                "Constructors": [
-                    {
-                        "constructorId": "mclaren",
-                        "url": "http://en.wikipedia.org/wiki/McLaren",
-                        "name": "McLaren",
-                        "nationality": "British",
-                    }
-                ],
-            },
-            {
-                "position": "10",
-                "positionText": "10",
-                "points": "8",
-                "wins": "0",
-                "Driver": {
-                    "driverId": "gasly",
-                    "permanentNumber": "10",
-                    "code": "GAS",
-                    "url": "http://en.wikipedia.org/wiki/Pierre_Gasly",
-                    "givenName": "Pierre",
-                    "familyName": "Gasly",
-                    "dateOfBirth": "1996-02-07",
-                    "nationality": "French",
-                },
-                "Constructors": [
-                    {
-                        "constructorId": "alpine",
-                        "url": "http://en.wikipedia.org/wiki/Alpine_F1_Team",
-                        "name": "Alpine F1 Team",
-                        "nationality": "French",
-                    }
-                ],
-            },
-            {
-                "position": "11",
-                "positionText": "11",
-                "points": "6",
-                "wins": "0",
-                "Driver": {
-                    "driverId": "hulkenberg",
-                    "permanentNumber": "27",
-                    "code": "HUL",
-                    "url": "http://en.wikipedia.org/wiki/Nico_H%C3%BClkenberg",
-                    "givenName": "Nico",
-                    "familyName": "Hülkenberg",
-                    "dateOfBirth": "1987-08-19",
-                    "nationality": "German",
-                },
-                "Constructors": [
-                    {
-                        "constructorId": "haas",
-                        "url": "http://en.wikipedia.org/wiki/Haas_F1_Team",
-                        "name": "Haas F1 Team",
-                        "nationality": "American",
-                    }
-                ],
-            },
-            {
-                "position": "12",
-                "positionText": "12",
-                "points": "6",
-                "wins": "0",
-                "Driver": {
-                    "driverId": "ocon",
-                    "permanentNumber": "31",
-                    "code": "OCO",
-                    "url": "http://en.wikipedia.org/wiki/Esteban_Ocon",
-                    "givenName": "Esteban",
-                    "familyName": "Ocon",
-                    "dateOfBirth": "1996-09-17",
-                    "nationality": "French",
-                },
-                "Constructors": [
-                    {
-                        "constructorId": "alpine",
-                        "url": "http://en.wikipedia.org/wiki/Alpine_F1_Team",
-                        "name": "Alpine F1 Team",
-                        "nationality": "French",
-                    }
-                ],
-            },
-            {
-                "position": "13",
-                "positionText": "13",
-                "points": "4",
-                "wins": "0",
-                "Driver": {
-                    "driverId": "bottas",
-                    "permanentNumber": "77",
-                    "code": "BOT",
-                    "url": "http://en.wikipedia.org/wiki/Valtteri_Bottas",
-                    "givenName": "Valtteri",
-                    "familyName": "Bottas",
-                    "dateOfBirth": "1989-08-28",
-                    "nationality": "Finnish",
-                },
-                "Constructors": [
-                    {
-                        "constructorId": "alfa",
-                        "url": "http://en.wikipedia.org/wiki/Alfa_Romeo_in_Formula_One",
-                        "name": "Alfa Romeo",
-                        "nationality": "Swiss",
-                    }
-                ],
-            },
-            {
-                "position": "14",
-                "positionText": "14",
-                "points": "4",
-                "wins": "0",
-                "Driver": {
-                    "driverId": "piastri",
-                    "permanentNumber": "81",
-                    "code": "PIA",
-                    "url": "http://en.wikipedia.org/wiki/Oscar_Piastri",
-                    "givenName": "Oscar",
-                    "familyName": "Piastri",
-                    "dateOfBirth": "2001-04-06",
-                    "nationality": "Australian",
-                },
-                "Constructors": [
-                    {
-                        "constructorId": "mclaren",
-                        "url": "http://en.wikipedia.org/wiki/McLaren",
-                        "name": "McLaren",
-                        "nationality": "British",
-                    }
-                ],
-            },
-            {
-                "position": "15",
-                "positionText": "15",
-                "points": "2",
-                "wins": "0",
-                "Driver": {
-                    "driverId": "zhou",
-                    "permanentNumber": "24",
-                    "code": "ZHO",
-                    "url": "http://en.wikipedia.org/wiki/Zhou_Guanyu",
-                    "givenName": "Guanyu",
-                    "familyName": "Zhou",
-                    "dateOfBirth": "1999-05-30",
-                    "nationality": "Chinese",
-                },
-                "Constructors": [
-                    {
-                        "constructorId": "alfa",
-                        "url": "http://en.wikipedia.org/wiki/Alfa_Romeo_in_Formula_One",
-                        "name": "Alfa Romeo",
-                        "nationality": "Swiss",
-                    }
-                ],
-            },
-            {
-                "position": "16",
-                "positionText": "16",
-                "points": "2",
-                "wins": "0",
-                "Driver": {
-                    "driverId": "tsunoda",
-                    "permanentNumber": "22",
-                    "code": "TSU",
-                    "url": "http://en.wikipedia.org/wiki/Yuki_Tsunoda",
-                    "givenName": "Yuki",
-                    "familyName": "Tsunoda",
-                    "dateOfBirth": "2000-05-11",
-                    "nationality": "Japanese",
-                },
-                "Constructors": [
-                    {
-                        "constructorId": "alphatauri",
-                        "url": "http://en.wikipedia.org/wiki/Scuderia_AlphaTauri",
-                        "name": "AlphaTauri",
-                        "nationality": "Italian",
-                    }
-                ],
-            },
-            {
-                "position": "17",
-                "positionText": "17",
-                "points": "2",
-                "wins": "0",
-                "Driver": {
-                    "driverId": "kevin_magnussen",
-                    "permanentNumber": "20",
-                    "code": "MAG",
-                    "url": "http://en.wikipedia.org/wiki/Kevin_Magnussen",
-                    "givenName": "Kevin",
-                    "familyName": "Magnussen",
-                    "dateOfBirth": "1992-10-05",
-                    "nationality": "Danish",
-                },
-                "Constructors": [
-                    {
-                        "constructorId": "haas",
-                        "url": "http://en.wikipedia.org/wiki/Haas_F1_Team",
-                        "name": "Haas F1 Team",
-                        "nationality": "American",
-                    }
-                ],
-            },
-            {
-                "position": "18",
-                "positionText": "18",
-                "points": "1",
-                "wins": "0",
-                "Driver": {
-                    "driverId": "albon",
-                    "permanentNumber": "23",
-                    "code": "ALB",
-                    "url": "http://en.wikipedia.org/wiki/Alexander_Albon",
-                    "givenName": "Alexander",
-                    "familyName": "Albon",
-                    "dateOfBirth": "1996-03-23",
-                    "nationality": "Thai",
-                },
-                "Constructors": [
-                    {
-                        "constructorId": "williams",
-                        "url": "http://en.wikipedia.org/wiki/Williams_Grand_Prix_Engineering",
-                        "name": "Williams",
-                        "nationality": "British",
-                    }
-                ],
-            },
-            {
-                "position": "19",
-                "positionText": "19",
-                "points": "0",
-                "wins": "0",
-                "Driver": {
-                    "driverId": "sargeant",
-                    "permanentNumber": "2",
-                    "code": "SAR",
-                    "url": "http://en.wikipedia.org/wiki/Logan_Sargeant",
-                    "givenName": "Logan",
-                    "familyName": "Sargeant",
-                    "dateOfBirth": "2000-12-31",
-                    "nationality": "American",
-                },
-                "Constructors": [
-                    {
-                        "constructorId": "williams",
-                        "url": "http://en.wikipedia.org/wiki/Williams_Grand_Prix_Engineering",
-                        "name": "Williams",
-                        "nationality": "British",
-                    }
-                ],
-            },
-            {
-                "position": "20",
-                "positionText": "20",
-                "points": "0",
-                "wins": "0",
-                "Driver": {
-                    "driverId": "de_vries",
-                    "permanentNumber": "21",
-                    "code": "DEV",
-                    "url": "http://en.wikipedia.org/wiki/Nyck_de_Vries",
-                    "givenName": "Nyck",
-                    "familyName": "de Vries",
-                    "dateOfBirth": "1995-02-06",
-                    "nationality": "Dutch",
-                },
-                "Constructors": [
-                    {
-                        "constructorId": "alphatauri",
-                        "url": "http://en.wikipedia.org/wiki/Scuderia_AlphaTauri",
-                        "name": "AlphaTauri",
-                        "nationality": "Italian",
-                    }
-                ],
-            },
-        ],
-        "ConstructorStandings": [
-            {
-                "position": "1",
-                "positionText": "1",
-                "points": "224",
-                "wins": "5",
-                "Constructor": {
-                    "constructorId": "red_bull",
-                    "url": "http://en.wikipedia.org/wiki/Red_Bull_Racing",
-                    "name": "Red Bull",
-                    "nationality": "Austrian",
-                },
-            },
-            {
-                "position": "2",
-                "positionText": "2",
-                "points": "102",
-                "wins": "0",
-                "Constructor": {
-                    "constructorId": "aston_martin",
-                    "url": "http://en.wikipedia.org/wiki/Aston_Martin_in_Formula_One",
-                    "name": "Aston Martin",
-                    "nationality": "British",
-                },
-            },
-            {
-                "position": "3",
-                "positionText": "3",
-                "points": "96",
-                "wins": "0",
-                "Constructor": {
-                    "constructorId": "mercedes",
-                    "url": "http://en.wikipedia.org/wiki/Mercedes-Benz_in_Formula_One",
-                    "name": "Mercedes",
-                    "nationality": "German",
-                },
-            },
-            {
-                "position": "4",
-                "positionText": "4",
-                "points": "78",
-                "wins": "0",
-                "Constructor": {
-                    "constructorId": "ferrari",
-                    "url": "http://en.wikipedia.org/wiki/Scuderia_Ferrari",
-                    "name": "Ferrari",
-                    "nationality": "Italian",
-                },
-            },
-            {
-                "position": "5",
-                "positionText": "5",
-                "points": "14",
-                "wins": "0",
-                "Constructor": {
-                    "constructorId": "mclaren",
-                    "url": "http://en.wikipedia.org/wiki/McLaren",
-                    "name": "McLaren",
-                    "nationality": "British",
-                },
-            },
-            {
-                "position": "6",
-                "positionText": "6",
-                "points": "14",
-                "wins": "0",
-                "Constructor": {
-                    "constructorId": "alpine",
-                    "url": "http://en.wikipedia.org/wiki/Alpine_F1_Team",
-                    "name": "Alpine F1 Team",
-                    "nationality": "French",
-                },
-            },
-            {
-                "position": "7",
-                "positionText": "7",
-                "points": "8",
-                "wins": "0",
-                "Constructor": {
-                    "constructorId": "haas",
-                    "url": "http://en.wikipedia.org/wiki/Haas_F1_Team",
-                    "name": "Haas F1 Team",
-                    "nationality": "American",
-                },
-            },
-            {
-                "position": "8",
-                "positionText": "8",
-                "points": "6",
-                "wins": "0",
-                "Constructor": {
-                    "constructorId": "alfa",
-                    "url": "http://en.wikipedia.org/wiki/Alfa_Romeo_in_Formula_One",
-                    "name": "Alfa Romeo",
-                    "nationality": "Swiss",
-                },
-            },
-            {
-                "position": "9",
-                "positionText": "9",
-                "points": "2",
-                "wins": "0",
-                "Constructor": {
-                    "constructorId": "alphatauri",
-                    "url": "http://en.wikipedia.org/wiki/Scuderia_AlphaTauri",
-                    "name": "AlphaTauri",
-                    "nationality": "Italian",
-                },
-            },
-            {
-                "position": "10",
-                "positionText": "10",
-                "points": "1",
-                "wins": "0",
-                "Constructor": {
-                    "constructorId": "williams",
-                    "url": "http://en.wikipedia.org/wiki/Williams_Grand_Prix_Engineering",
-                    "name": "Williams",
-                    "nationality": "British",
-                },
-            },
-        ],
-    }
-
-
-# endregion standings - good inputs
-
-# region standings - no inputs
-
-
-def test_get_standings_good_round_bad_year_no_input():
-    response = client.get("/standings?round=3")
-    assert response.status_code == status.HTTP_400_BAD_REQUEST
-    assert response.json() == {"detail": 'Bad request. Must provide the "year" parameter.'}
-
-
-# endregion standings - no inputs
-
-# endregion standings
-
-
-# region results
-
-# region results - good inputs
-
-
-def test_get_results():
-    response = client.get("/results/2023/5")
-    assert response.status_code == status.HTTP_200_OK
-    assert response.json() == [
-        {
-            "DriverNumber": "1",
-            "BroadcastName": "M VERSTAPPEN",
-            "Abbreviation": "VER",
-            "DriverId": "max_verstappen",
-            "TeamName": "Red Bull Racing",
-            "TeamColor": "3671C6",
-            "TeamId": "red_bull",
-            "FirstName": "Max",
-            "LastName": "Verstappen",
-            "FullName": "Max Verstappen",
-            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/M/MAXVER01_Max_Verstappen/maxver01.png.transform/1col/image.png",
-            "CountryCode": "NED",
-            "Position": 1,
-            "ClassifiedPosition": "1",
-            "GridPosition": 9,
-            "Q1": None,
-            "Q2": None,
-            "Q3": None,
-            "Time": 5258241,
-            "Status": "Finished",
-            "Points": 26,
-        },
-        {
-            "DriverNumber": "11",
-            "BroadcastName": "S PEREZ",
-            "Abbreviation": "PER",
-            "DriverId": "perez",
-            "TeamName": "Red Bull Racing",
-            "TeamColor": "3671C6",
-            "TeamId": "red_bull",
-            "FirstName": "Sergio",
-            "LastName": "Perez",
-            "FullName": "Sergio Perez",
-            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/S/SERPER01_Sergio_Perez/serper01.png.transform/1col/image.png",
-            "CountryCode": "MEX",
-            "Position": 2,
-            "ClassifiedPosition": "2",
-            "GridPosition": 1,
-            "Q1": None,
-            "Q2": None,
-            "Q3": None,
-            "Time": 5384,
-            "Status": "Finished",
-            "Points": 18,
-        },
-        {
-            "DriverNumber": "14",
-            "BroadcastName": "F ALONSO",
-            "Abbreviation": "ALO",
-            "DriverId": "alonso",
-            "TeamName": "Aston Martin",
-            "TeamColor": "358C75",
-            "TeamId": "aston_martin",
-            "FirstName": "Fernando",
-            "LastName": "Alonso",
-            "FullName": "Fernando Alonso",
-            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/F/FERALO01_Fernando_Alonso/feralo01.png.transform/1col/image.png",
-            "CountryCode": "ESP",
-            "Position": 3,
-            "ClassifiedPosition": "3",
-            "GridPosition": 2,
-            "Q1": None,
-            "Q2": None,
-            "Q3": None,
-            "Time": 26305,
-            "Status": "Finished",
-            "Points": 15,
-        },
-        {
-            "DriverNumber": "63",
-            "BroadcastName": "G RUSSELL",
-            "Abbreviation": "RUS",
-            "DriverId": "russell",
-            "TeamName": "Mercedes",
-            "TeamColor": "6CD3BF",
-            "TeamId": "mercedes",
-            "FirstName": "George",
-            "LastName": "Russell",
-            "FullName": "George Russell",
-            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/G/GEORUS01_George_Russell/georus01.png.transform/1col/image.png",
-            "CountryCode": "GBR",
-            "Position": 4,
-            "ClassifiedPosition": "4",
-            "GridPosition": 6,
-            "Q1": None,
-            "Q2": None,
-            "Q3": None,
-            "Time": 33229,
-            "Status": "Finished",
-            "Points": 12,
-        },
-        {
-            "DriverNumber": "55",
-            "BroadcastName": "C SAINZ",
-            "Abbreviation": "SAI",
-            "DriverId": "sainz",
-            "TeamName": "Ferrari",
-            "TeamColor": "F91536",
-            "TeamId": "ferrari",
-            "FirstName": "Carlos",
-            "LastName": "Sainz",
-            "FullName": "Carlos Sainz",
-            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/C/CARSAI01_Carlos_Sainz/carsai01.png.transform/1col/image.png",
-            "CountryCode": "ESP",
-            "Position": 5,
-            "ClassifiedPosition": "5",
-            "GridPosition": 3,
-            "Q1": None,
-            "Q2": None,
-            "Q3": None,
-            "Time": 42511,
-            "Status": "Finished",
-            "Points": 10,
-        },
-        {
-            "DriverNumber": "44",
-            "BroadcastName": "L HAMILTON",
-            "Abbreviation": "HAM",
-            "DriverId": "hamilton",
-            "TeamName": "Mercedes",
-            "TeamColor": "6CD3BF",
-            "TeamId": "mercedes",
-            "FirstName": "Lewis",
-            "LastName": "Hamilton",
-            "FullName": "Lewis Hamilton",
-            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/L/LEWHAM01_Lewis_Hamilton/lewham01.png.transform/1col/image.png",
-            "CountryCode": "GBR",
-            "Position": 6,
-            "ClassifiedPosition": "6",
-            "GridPosition": 13,
-            "Q1": None,
-            "Q2": None,
-            "Q3": None,
-            "Time": 51249,
-            "Status": "Finished",
-            "Points": 8,
-        },
-        {
-            "DriverNumber": "16",
-            "BroadcastName": "C LECLERC",
-            "Abbreviation": "LEC",
-            "DriverId": "leclerc",
-            "TeamName": "Ferrari",
-            "TeamColor": "F91536",
-            "TeamId": "ferrari",
-            "FirstName": "Charles",
-            "LastName": "Leclerc",
-            "FullName": "Charles Leclerc",
-            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/C/CHALEC01_Charles_Leclerc/chalec01.png.transform/1col/image.png",
-            "CountryCode": "MON",
-            "Position": 7,
-            "ClassifiedPosition": "7",
-            "GridPosition": 7,
-            "Q1": None,
-            "Q2": None,
-            "Q3": None,
-            "Time": 52988,
-            "Status": "Finished",
-            "Points": 6,
-        },
-        {
-            "DriverNumber": "10",
-            "BroadcastName": "P GASLY",
-            "Abbreviation": "GAS",
-            "DriverId": "gasly",
-            "TeamName": "Alpine",
-            "TeamColor": "2293D1",
-            "TeamId": "alpine",
-            "FirstName": "Pierre",
-            "LastName": "Gasly",
-            "FullName": "Pierre Gasly",
-            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/P/PIEGAS01_Pierre_Gasly/piegas01.png.transform/1col/image.png",
-            "CountryCode": "FRA",
-            "Position": 8,
-            "ClassifiedPosition": "8",
-            "GridPosition": 5,
-            "Q1": None,
-            "Q2": None,
-            "Q3": None,
-            "Time": 55670,
-            "Status": "Finished",
-            "Points": 4,
-        },
-        {
-            "DriverNumber": "31",
-            "BroadcastName": "E OCON",
-            "Abbreviation": "OCO",
-            "DriverId": "ocon",
-            "TeamName": "Alpine",
-            "TeamColor": "2293D1",
-            "TeamId": "alpine",
-            "FirstName": "Esteban",
-            "LastName": "Ocon",
-            "FullName": "Esteban Ocon",
-            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/E/ESTOCO01_Esteban_Ocon/estoco01.png.transform/1col/image.png",
-            "CountryCode": "FRA",
-            "Position": 9,
-            "ClassifiedPosition": "9",
-            "GridPosition": 8,
-            "Q1": None,
-            "Q2": None,
-            "Q3": None,
-            "Time": 58123,
-            "Status": "Finished",
-            "Points": 2,
-        },
-        {
-            "DriverNumber": "20",
-            "BroadcastName": "K MAGNUSSEN",
-            "Abbreviation": "MAG",
-            "DriverId": "kevin_magnussen",
-            "TeamName": "Haas F1 Team",
-            "TeamColor": "B6BABD",
-            "TeamId": "haas",
-            "FirstName": "Kevin",
-            "LastName": "Magnussen",
-            "FullName": "Kevin Magnussen",
-            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/K/KEVMAG01_Kevin_Magnussen/kevmag01.png.transform/1col/image.png",
-            "CountryCode": "DEN",
-            "Position": 10,
-            "ClassifiedPosition": "10",
-            "GridPosition": 4,
-            "Q1": None,
-            "Q2": None,
-            "Q3": None,
-            "Time": 62945,
-            "Status": "Finished",
-            "Points": 1,
-        },
-        {
-            "DriverNumber": "22",
-            "BroadcastName": "Y TSUNODA",
-            "Abbreviation": "TSU",
-            "DriverId": "tsunoda",
-            "TeamName": "AlphaTauri",
-            "TeamColor": "5E8FAA",
-            "TeamId": "alphatauri",
-            "FirstName": "Yuki",
-            "LastName": "Tsunoda",
-            "FullName": "Yuki Tsunoda",
-            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/Y/YUKTSU01_Yuki_Tsunoda/yuktsu01.png.transform/1col/image.png",
-            "CountryCode": "JPN",
-            "Position": 11,
-            "ClassifiedPosition": "11",
-            "GridPosition": 17,
-            "Q1": None,
-            "Q2": None,
-            "Q3": None,
-            "Time": 64309,
-            "Status": "Finished",
-            "Points": 0,
-        },
-        {
-            "DriverNumber": "18",
-            "BroadcastName": "L STROLL",
-            "Abbreviation": "STR",
-            "DriverId": "stroll",
-            "TeamName": "Aston Martin",
-            "TeamColor": "358C75",
-            "TeamId": "aston_martin",
-            "FirstName": "Lance",
-            "LastName": "Stroll",
-            "FullName": "Lance Stroll",
-            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/L/LANSTR01_Lance_Stroll/lanstr01.png.transform/1col/image.png",
-            "CountryCode": "CAN",
-            "Position": 12,
-            "ClassifiedPosition": "12",
-            "GridPosition": 18,
-            "Q1": None,
-            "Q2": None,
-            "Q3": None,
-            "Time": 64754,
-            "Status": "Finished",
-            "Points": 0,
-        },
-        {
-            "DriverNumber": "77",
-            "BroadcastName": "V BOTTAS",
-            "Abbreviation": "BOT",
-            "DriverId": "bottas",
-            "TeamName": "Alfa Romeo",
-            "TeamColor": "C92D4B",
-            "TeamId": "alfa",
-            "FirstName": "Valtteri",
-            "LastName": "Bottas",
-            "FullName": "Valtteri Bottas",
-            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/V/VALBOT01_Valtteri_Bottas/valbot01.png.transform/1col/image.png",
-            "CountryCode": "FIN",
-            "Position": 13,
-            "ClassifiedPosition": "13",
-            "GridPosition": 10,
-            "Q1": None,
-            "Q2": None,
-            "Q3": None,
-            "Time": 71637,
-            "Status": "Finished",
-            "Points": 0,
-        },
-        {
-            "DriverNumber": "23",
-            "BroadcastName": "A ALBON",
-            "Abbreviation": "ALB",
-            "DriverId": "albon",
-            "TeamName": "Williams",
-            "TeamColor": "37BEDD",
-            "TeamId": "williams",
-            "FirstName": "Alexander",
-            "LastName": "Albon",
-            "FullName": "Alexander Albon",
-            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/A/ALEALB01_Alexander_Albon/alealb01.png.transform/1col/image.png",
-            "CountryCode": "THA",
-            "Position": 14,
-            "ClassifiedPosition": "14",
-            "GridPosition": 11,
-            "Q1": None,
-            "Q2": None,
-            "Q3": None,
-            "Time": 72861,
-            "Status": "Finished",
-            "Points": 0,
-        },
-        {
-            "DriverNumber": "27",
-            "BroadcastName": "N HULKENBERG",
-            "Abbreviation": "HUL",
-            "DriverId": "hulkenberg",
-            "TeamName": "Haas F1 Team",
-            "TeamColor": "B6BABD",
-            "TeamId": "haas",
-            "FirstName": "Nico",
-            "LastName": "Hulkenberg",
-            "FullName": "Nico Hulkenberg",
-            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/N/NICHUL01_Nico_Hulkenberg/nichul01.png.transform/1col/image.png",
-            "CountryCode": "GER",
-            "Position": 15,
-            "ClassifiedPosition": "15",
-            "GridPosition": 12,
-            "Q1": None,
-            "Q2": None,
-            "Q3": None,
-            "Time": 74950,
-            "Status": "Finished",
-            "Points": 0,
-        },
-        {
-            "DriverNumber": "24",
-            "BroadcastName": "G ZHOU",
-            "Abbreviation": "ZHO",
-            "DriverId": "zhou",
-            "TeamName": "Alfa Romeo",
-            "TeamColor": "C92D4B",
-            "TeamId": "alfa",
-            "FirstName": "Guanyu",
-            "LastName": "Zhou",
-            "FullName": "Guanyu Zhou",
-            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/G/GUAZHO01_Guanyu_Zhou/guazho01.png.transform/1col/image.png",
-            "CountryCode": "CHN",
-            "Position": 16,
-            "ClassifiedPosition": "16",
-            "GridPosition": 14,
-            "Q1": None,
-            "Q2": None,
-            "Q3": None,
-            "Time": 78440,
-            "Status": "Finished",
-            "Points": 0,
-        },
-        {
-            "DriverNumber": "4",
-            "BroadcastName": "L NORRIS",
-            "Abbreviation": "NOR",
-            "DriverId": "norris",
-            "TeamName": "McLaren",
-            "TeamColor": "F58020",
-            "TeamId": "mclaren",
-            "FirstName": "Lando",
-            "LastName": "Norris",
-            "FullName": "Lando Norris",
-            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/L/LANNOR01_Lando_Norris/lannor01.png.transform/1col/image.png",
-            "CountryCode": "GBR",
-            "Position": 17,
-            "ClassifiedPosition": "17",
-            "GridPosition": 16,
-            "Q1": None,
-            "Q2": None,
-            "Q3": None,
-            "Time": 87717,
-            "Status": "Finished",
-            "Points": 0,
-        },
-        {
-            "DriverNumber": "21",
-            "BroadcastName": "N DE VRIES",
-            "Abbreviation": "DEV",
-            "DriverId": "de_vries",
-            "TeamName": "AlphaTauri",
-            "TeamColor": "5E8FAA",
-            "TeamId": "alphatauri",
-            "FirstName": "Nyck",
-            "LastName": "De Vries",
-            "FullName": "Nyck De Vries",
-            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/N/NYCDEV01_Nyck_De%20Vries/nycdev01.png.transform/1col/image.png",
-            "CountryCode": "NED",
-            "Position": 18,
-            "ClassifiedPosition": "18",
-            "GridPosition": 15,
-            "Q1": None,
-            "Q2": None,
-            "Q3": None,
-            "Time": 88949,
-            "Status": "Finished",
-            "Points": 0,
-        },
-        {
-            "DriverNumber": "81",
-            "BroadcastName": "O PIASTRI",
-            "Abbreviation": "PIA",
-            "DriverId": "piastri",
-            "TeamName": "McLaren",
-            "TeamColor": "F58020",
-            "TeamId": "mclaren",
-            "FirstName": "Oscar",
-            "LastName": "Piastri",
-            "FullName": "Oscar Piastri",
-            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/O/OSCPIA01_Oscar_Piastri/oscpia01.png.transform/1col/image.png",
-            "CountryCode": "AUS",
-            "Position": 19,
-            "ClassifiedPosition": "19",
-            "GridPosition": 19,
-            "Q1": None,
-            "Q2": None,
-            "Q3": None,
-            "Time": None,
-            "Status": "+1 Lap",
-            "Points": 0,
-        },
-        {
-            "DriverNumber": "2",
-            "BroadcastName": "L SARGEANT",
-            "Abbreviation": "SAR",
-            "DriverId": "sargeant",
-            "TeamName": "Williams",
-            "TeamColor": "37BEDD",
-            "TeamId": "williams",
-            "FirstName": "Logan",
-            "LastName": "Sargeant",
-            "FullName": "Logan Sargeant",
-            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/L/LOGSAR01_Logan_Sargeant/logsar01.png.transform/1col/image.png",
-            "CountryCode": "USA",
-            "Position": 20,
-            "ClassifiedPosition": "20",
-            "GridPosition": 20,
-            "Q1": None,
-            "Q2": None,
-            "Q3": None,
-            "Time": None,
-            "Status": "+1 Lap",
-            "Points": 0,
-        },
-    ]
-
-
-def test_get_results_with_session():
-    response = client.get("/results/2023/5?session=4")
-    assert response.status_code == status.HTTP_200_OK
-    assert response.json() == [
-        {
-            "DriverNumber": "11",
-            "BroadcastName": "S PEREZ",
-            "Abbreviation": "PER",
-            "DriverId": "perez",
-            "TeamName": "Red Bull Racing",
-            "TeamColor": "3671C6",
-            "TeamId": "red_bull",
-            "FirstName": "Sergio",
-            "LastName": "Perez",
-            "FullName": "Sergio Perez",
-            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/S/SERPER01_Sergio_Perez/serper01.png.transform/1col/image.png",
-            "CountryCode": "MEX",
-            "Position": 1,
-            "ClassifiedPosition": "",
-            "GridPosition": None,
-            "Q1": 87713,
-            "Q2": 87328,
-            "Q3": 86841,
-            "Time": None,
-            "Status": "",
-            "Points": None,
-        },
-        {
-            "DriverNumber": "14",
-            "BroadcastName": "F ALONSO",
-            "Abbreviation": "ALO",
-            "DriverId": "alonso",
-            "TeamName": "Aston Martin",
-            "TeamColor": "358C75",
-            "TeamId": "aston_martin",
-            "FirstName": "Fernando",
-            "LastName": "Alonso",
-            "FullName": "Fernando Alonso",
-            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/F/FERALO01_Fernando_Alonso/feralo01.png.transform/1col/image.png",
-            "CountryCode": "ESP",
-            "Position": 2,
-            "ClassifiedPosition": "",
-            "GridPosition": None,
-            "Q1": 88179,
-            "Q2": 87097,
-            "Q3": 87202,
-            "Time": None,
-            "Status": "",
-            "Points": None,
-        },
-        {
-            "DriverNumber": "55",
-            "BroadcastName": "C SAINZ",
-            "Abbreviation": "SAI",
-            "DriverId": "sainz",
-            "TeamName": "Ferrari",
-            "TeamColor": "F91536",
-            "TeamId": "ferrari",
-            "FirstName": "Carlos",
-            "LastName": "Sainz",
-            "FullName": "Carlos Sainz",
-            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/C/CARSAI01_Carlos_Sainz/carsai01.png.transform/1col/image.png",
-            "CountryCode": "ESP",
-            "Position": 3,
-            "ClassifiedPosition": "",
-            "GridPosition": None,
-            "Q1": 87686,
-            "Q2": 87148,
-            "Q3": 87349,
-            "Time": None,
-            "Status": "",
-            "Points": None,
-        },
-        {
-            "DriverNumber": "20",
-            "BroadcastName": "K MAGNUSSEN",
-            "Abbreviation": "MAG",
-            "DriverId": "kevin_magnussen",
-            "TeamName": "Haas F1 Team",
-            "TeamColor": "B6BABD",
-            "TeamId": "haas",
-            "FirstName": "Kevin",
-            "LastName": "Magnussen",
-            "FullName": "Kevin Magnussen",
-            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/K/KEVMAG01_Kevin_Magnussen/kevmag01.png.transform/1col/image.png",
-            "CountryCode": "DEN",
-            "Position": 4,
-            "ClassifiedPosition": "",
-            "GridPosition": None,
-            "Q1": 87809,
-            "Q2": 87673,
-            "Q3": 87767,
-            "Time": None,
-            "Status": "",
-            "Points": None,
-        },
-        {
-            "DriverNumber": "10",
-            "BroadcastName": "P GASLY",
-            "Abbreviation": "GAS",
-            "DriverId": "gasly",
-            "TeamName": "Alpine",
-            "TeamColor": "2293D1",
-            "TeamId": "alpine",
-            "FirstName": "Pierre",
-            "LastName": "Gasly",
-            "FullName": "Pierre Gasly",
-            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/P/PIEGAS01_Pierre_Gasly/piegas01.png.transform/1col/image.png",
-            "CountryCode": "FRA",
-            "Position": 5,
-            "ClassifiedPosition": "",
-            "GridPosition": None,
-            "Q1": 88061,
-            "Q2": 87612,
-            "Q3": 87786,
-            "Time": None,
-            "Status": "",
-            "Points": None,
-        },
-        {
-            "DriverNumber": "63",
-            "BroadcastName": "G RUSSELL",
-            "Abbreviation": "RUS",
-            "DriverId": "russell",
-            "TeamName": "Mercedes",
-            "TeamColor": "6CD3BF",
-            "TeamId": "mercedes",
-            "FirstName": "George",
-            "LastName": "Russell",
-            "FullName": "George Russell",
-            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/G/GEORUS01_George_Russell/georus01.png.transform/1col/image.png",
-            "CountryCode": "GBR",
-            "Position": 6,
-            "ClassifiedPosition": "",
-            "GridPosition": None,
-            "Q1": 88086,
-            "Q2": 87743,
-            "Q3": 87804,
-            "Time": None,
-            "Status": "",
-            "Points": None,
-        },
-        {
-            "DriverNumber": "16",
-            "BroadcastName": "C LECLERC",
-            "Abbreviation": "LEC",
-            "DriverId": "leclerc",
-            "TeamName": "Ferrari",
-            "TeamColor": "F91536",
-            "TeamId": "ferrari",
-            "FirstName": "Charles",
-            "LastName": "Leclerc",
-            "FullName": "Charles Leclerc",
-            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/C/CHALEC01_Charles_Leclerc/chalec01.png.transform/1col/image.png",
-            "CountryCode": "MON",
-            "Position": 7,
-            "ClassifiedPosition": "",
-            "GridPosition": None,
-            "Q1": 87713,
-            "Q2": 86964,
-            "Q3": 87861,
-            "Time": None,
-            "Status": "",
-            "Points": None,
-        },
-        {
-            "DriverNumber": "31",
-            "BroadcastName": "E OCON",
-            "Abbreviation": "OCO",
-            "DriverId": "ocon",
-            "TeamName": "Alpine",
-            "TeamColor": "2293D1",
-            "TeamId": "alpine",
-            "FirstName": "Esteban",
-            "LastName": "Ocon",
-            "FullName": "Esteban Ocon",
-            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/E/ESTOCO01_Esteban_Ocon/estoco01.png.transform/1col/image.png",
-            "CountryCode": "FRA",
-            "Position": 8,
-            "ClassifiedPosition": "",
-            "GridPosition": None,
-            "Q1": 87872,
-            "Q2": 87444,
-            "Q3": 87935,
-            "Time": None,
-            "Status": "",
-            "Points": None,
-        },
-        {
-            "DriverNumber": "1",
-            "BroadcastName": "M VERSTAPPEN",
-            "Abbreviation": "VER",
-            "DriverId": "max_verstappen",
-            "TeamName": "Red Bull Racing",
-            "TeamColor": "3671C6",
-            "TeamId": "red_bull",
-            "FirstName": "Max",
-            "LastName": "Verstappen",
-            "FullName": "Max Verstappen",
-            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/M/MAXVER01_Max_Verstappen/maxver01.png.transform/1col/image.png",
-            "CountryCode": "NED",
-            "Position": 9,
-            "ClassifiedPosition": "",
-            "GridPosition": None,
-            "Q1": 87363,
-            "Q2": 86814,
-            "Q3": None,
-            "Time": None,
-            "Status": "",
-            "Points": None,
-        },
-        {
-            "DriverNumber": "77",
-            "BroadcastName": "V BOTTAS",
-            "Abbreviation": "BOT",
-            "DriverId": "bottas",
-            "TeamName": "Alfa Romeo",
-            "TeamColor": "C92D4B",
-            "TeamId": "alfa",
-            "FirstName": "Valtteri",
-            "LastName": "Bottas",
-            "FullName": "Valtteri Bottas",
-            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/V/VALBOT01_Valtteri_Bottas/valbot01.png.transform/1col/image.png",
-            "CountryCode": "FIN",
-            "Position": 10,
-            "ClassifiedPosition": "",
-            "GridPosition": None,
-            "Q1": 87864,
-            "Q2": 87564,
-            "Q3": None,
-            "Time": None,
-            "Status": "",
-            "Points": None,
-        },
-        {
-            "DriverNumber": "23",
-            "BroadcastName": "A ALBON",
-            "Abbreviation": "ALB",
-            "DriverId": "albon",
-            "TeamName": "Williams",
-            "TeamColor": "37BEDD",
-            "TeamId": "williams",
-            "FirstName": "Alexander",
-            "LastName": "Albon",
-            "FullName": "Alexander Albon",
-            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/A/ALEALB01_Alexander_Albon/alealb01.png.transform/1col/image.png",
-            "CountryCode": "THA",
-            "Position": 11,
-            "ClassifiedPosition": "",
-            "GridPosition": None,
-            "Q1": 88234,
-            "Q2": 87795,
-            "Q3": None,
-            "Time": None,
-            "Status": "",
-            "Points": None,
-        },
-        {
-            "DriverNumber": "27",
-            "BroadcastName": "N HULKENBERG",
-            "Abbreviation": "HUL",
-            "DriverId": "hulkenberg",
-            "TeamName": "Haas F1 Team",
-            "TeamColor": "B6BABD",
-            "TeamId": "haas",
-            "FirstName": "Nico",
-            "LastName": "Hulkenberg",
-            "FullName": "Nico Hulkenberg",
-            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/N/NICHUL01_Nico_Hulkenberg/nichul01.png.transform/1col/image.png",
-            "CountryCode": "GER",
-            "Position": 12,
-            "ClassifiedPosition": "",
-            "GridPosition": None,
-            "Q1": 87945,
-            "Q2": 87903,
-            "Q3": None,
-            "Time": None,
-            "Status": "",
-            "Points": None,
-        },
-        {
-            "DriverNumber": "44",
-            "BroadcastName": "L HAMILTON",
-            "Abbreviation": "HAM",
-            "DriverId": "hamilton",
-            "TeamName": "Mercedes",
-            "TeamColor": "6CD3BF",
-            "TeamId": "mercedes",
-            "FirstName": "Lewis",
-            "LastName": "Hamilton",
-            "FullName": "Lewis Hamilton",
-            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/L/LEWHAM01_Lewis_Hamilton/lewham01.png.transform/1col/image.png",
-            "CountryCode": "GBR",
-            "Position": 13,
-            "ClassifiedPosition": "",
-            "GridPosition": None,
-            "Q1": 87846,
-            "Q2": 87975,
-            "Q3": None,
-            "Time": None,
-            "Status": "",
-            "Points": None,
-        },
-        {
-            "DriverNumber": "24",
-            "BroadcastName": "G ZHOU",
-            "Abbreviation": "ZHO",
-            "DriverId": "zhou",
-            "TeamName": "Alfa Romeo",
-            "TeamColor": "C92D4B",
-            "TeamId": "alfa",
-            "FirstName": "Guanyu",
-            "LastName": "Zhou",
-            "FullName": "Guanyu Zhou",
-            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/G/GUAZHO01_Guanyu_Zhou/guazho01.png.transform/1col/image.png",
-            "CountryCode": "CHN",
-            "Position": 14,
-            "ClassifiedPosition": "",
-            "GridPosition": None,
-            "Q1": 88180,
-            "Q2": 88091,
-            "Q3": None,
-            "Time": None,
-            "Status": "",
-            "Points": None,
-        },
-        {
-            "DriverNumber": "21",
-            "BroadcastName": "N DE VRIES",
-            "Abbreviation": "DEV",
-            "DriverId": "de_vries",
-            "TeamName": "AlphaTauri",
-            "TeamColor": "5E8FAA",
-            "TeamId": "alphatauri",
-            "FirstName": "Nyck",
-            "LastName": "De Vries",
-            "FullName": "Nyck De Vries",
-            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/N/NYCDEV01_Nyck_De%20Vries/nycdev01.png.transform/1col/image.png",
-            "CountryCode": "NED",
-            "Position": 15,
-            "ClassifiedPosition": "",
-            "GridPosition": None,
-            "Q1": 88325,
-            "Q2": 88395,
-            "Q3": None,
-            "Time": None,
-            "Status": "",
-            "Points": None,
-        },
-        {
-            "DriverNumber": "4",
-            "BroadcastName": "L NORRIS",
-            "Abbreviation": "NOR",
-            "DriverId": "norris",
-            "TeamName": "McLaren",
-            "TeamColor": "F58020",
-            "TeamId": "mclaren",
-            "FirstName": "Lando",
-            "LastName": "Norris",
-            "FullName": "Lando Norris",
-            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/L/LANNOR01_Lando_Norris/lannor01.png.transform/1col/image.png",
-            "CountryCode": "GBR",
-            "Position": 16,
-            "ClassifiedPosition": "",
-            "GridPosition": None,
-            "Q1": 88394,
-            "Q2": None,
-            "Q3": None,
-            "Time": None,
-            "Status": "",
-            "Points": None,
-        },
-        {
-            "DriverNumber": "22",
-            "BroadcastName": "Y TSUNODA",
-            "Abbreviation": "TSU",
-            "DriverId": "tsunoda",
-            "TeamName": "AlphaTauri",
-            "TeamColor": "5E8FAA",
-            "TeamId": "alphatauri",
-            "FirstName": "Yuki",
-            "LastName": "Tsunoda",
-            "FullName": "Yuki Tsunoda",
-            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/Y/YUKTSU01_Yuki_Tsunoda/yuktsu01.png.transform/1col/image.png",
-            "CountryCode": "JPN",
-            "Position": 17,
-            "ClassifiedPosition": "",
-            "GridPosition": None,
-            "Q1": 88429,
-            "Q2": None,
-            "Q3": None,
-            "Time": None,
-            "Status": "",
-            "Points": None,
-        },
-        {
-            "DriverNumber": "18",
-            "BroadcastName": "L STROLL",
-            "Abbreviation": "STR",
-            "DriverId": "stroll",
-            "TeamName": "Aston Martin",
-            "TeamColor": "358C75",
-            "TeamId": "aston_martin",
-            "FirstName": "Lance",
-            "LastName": "Stroll",
-            "FullName": "Lance Stroll",
-            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/L/LANSTR01_Lance_Stroll/lanstr01.png.transform/1col/image.png",
-            "CountryCode": "CAN",
-            "Position": 18,
-            "ClassifiedPosition": "",
-            "GridPosition": None,
-            "Q1": 88476,
-            "Q2": None,
-            "Q3": None,
-            "Time": None,
-            "Status": "",
-            "Points": None,
-        },
-        {
-            "DriverNumber": "81",
-            "BroadcastName": "O PIASTRI",
-            "Abbreviation": "PIA",
-            "DriverId": "piastri",
-            "TeamName": "McLaren",
-            "TeamColor": "F58020",
-            "TeamId": "mclaren",
-            "FirstName": "Oscar",
-            "LastName": "Piastri",
-            "FullName": "Oscar Piastri",
-            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/O/OSCPIA01_Oscar_Piastri/oscpia01.png.transform/1col/image.png",
-            "CountryCode": "AUS",
-            "Position": 19,
-            "ClassifiedPosition": "",
-            "GridPosition": None,
-            "Q1": 88484,
-            "Q2": None,
-            "Q3": None,
-            "Time": None,
-            "Status": "",
-            "Points": None,
-        },
-        {
-            "DriverNumber": "2",
-            "BroadcastName": "L SARGEANT",
-            "Abbreviation": "SAR",
-            "DriverId": "sargeant",
-            "TeamName": "Williams",
-            "TeamColor": "37BEDD",
-            "TeamId": "williams",
-            "FirstName": "Logan",
-            "LastName": "Sargeant",
-            "FullName": "Logan Sargeant",
-            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/L/LOGSAR01_Logan_Sargeant/logsar01.png.transform/1col/image.png",
-            "CountryCode": "USA",
-            "Position": 20,
-            "ClassifiedPosition": "",
-            "GridPosition": None,
-            "Q1": 88577,
-            "Q2": None,
-            "Q3": None,
-            "Time": None,
-            "Status": "",
-            "Points": None,
-        },
-    ]
-
-
-# endregion results - good inputs
-
-# region results - bad inputs
-
-
-def test_get_results_bad_round_invalid():
-    response = client.get("/results/2023/25?session=2")
-    assert response.status_code == status.HTTP_400_BAD_REQUEST
-    assert response.json() == {"detail": "Bad Request. Invalid round: 25"}
-
-
-# endregion results - bad inputs
-
-# endregion results
-
-
-# region laps
-
-# region laps - good inputs
-
-
-def test_get_laps():
-    response = client.get("/laps/2023/5")
-    assert response.status_code == status.HTTP_200_OK
-    assert response.json() == []
-
-
-def test_get_laps_with_session_only():
-    response = client.get("/laps/2023/5?session=4")
-    assert response.status_code == status.HTTP_200_OK
-    assert response.json() == []
-
-
-def test_get_laps_with_driver_numbers_only():
-    response = client.get("/laps/2023/5?driver_numbers=1,44")
-    assert response.status_code == status.HTTP_200_OK
-    assert response.json() == []
-
-
-def test_get_laps_with_session_and_driver_numbers():
-    response = client.get("/laps/2023/5?session=4?driver_numbers=1,44")
-    assert response.status_code == status.HTTP_200_OK
-    assert response.json() == []
-
-
-# endregion laps - good inputs
-
-# region laps - bad inputs
-
-
-def test_get_laps_bad_driver_numbers():
-    response = client.get("/laps/2023/5?driver_numbers=0,1,13,44,83")
-    assert response.status_code == status.HTTP_200_OK
-    assert response.json() == []
-
-
-# endregion laps - bad inputs
-
-# endregion laps
