+++ tests/test_results.py
@@ -1,0 +1,964 @@
+# External
+from fastapi import status
+from fastapi.testclient import TestClient
+
+# Project
+from app.main import app
+
+
+client = TestClient(app)
+
+
+# region results
+
+# region results - good inputs
+
+
+def test_get_results():
+    response = client.get("/results/2023/5")
+    assert response.status_code == status.HTTP_200_OK
+    assert response.json() == [
+        {
+            "DriverNumber": "1",
+            "BroadcastName": "M VERSTAPPEN",
+            "Abbreviation": "VER",
+            "DriverId": "max_verstappen",
+            "TeamName": "Red Bull Racing",
+            "TeamColor": "3671C6",
+            "TeamId": "red_bull",
+            "FirstName": "Max",
+            "LastName": "Verstappen",
+            "FullName": "Max Verstappen",
+            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/M/MAXVER01_Max_Verstappen/maxver01.png.transform/1col/image.png",
+            "CountryCode": "NED",
+            "Position": 1,
+            "ClassifiedPosition": "1",
+            "GridPosition": 9,
+            "Q1": None,
+            "Q2": None,
+            "Q3": None,
+            "Time": 5258241,
+            "Status": "Finished",
+            "Points": 26,
+        },
+        {
+            "DriverNumber": "11",
+            "BroadcastName": "S PEREZ",
+            "Abbreviation": "PER",
+            "DriverId": "perez",
+            "TeamName": "Red Bull Racing",
+            "TeamColor": "3671C6",
+            "TeamId": "red_bull",
+            "FirstName": "Sergio",
+            "LastName": "Perez",
+            "FullName": "Sergio Perez",
+            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/S/SERPER01_Sergio_Perez/serper01.png.transform/1col/image.png",
+            "CountryCode": "MEX",
+            "Position": 2,
+            "ClassifiedPosition": "2",
+            "GridPosition": 1,
+            "Q1": None,
+            "Q2": None,
+            "Q3": None,
+            "Time": 5384,
+            "Status": "Finished",
+            "Points": 18,
+        },
+        {
+            "DriverNumber": "14",
+            "BroadcastName": "F ALONSO",
+            "Abbreviation": "ALO",
+            "DriverId": "alonso",
+            "TeamName": "Aston Martin",
+            "TeamColor": "358C75",
+            "TeamId": "aston_martin",
+            "FirstName": "Fernando",
+            "LastName": "Alonso",
+            "FullName": "Fernando Alonso",
+            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/F/FERALO01_Fernando_Alonso/feralo01.png.transform/1col/image.png",
+            "CountryCode": "ESP",
+            "Position": 3,
+            "ClassifiedPosition": "3",
+            "GridPosition": 2,
+            "Q1": None,
+            "Q2": None,
+            "Q3": None,
+            "Time": 26305,
+            "Status": "Finished",
+            "Points": 15,
+        },
+        {
+            "DriverNumber": "63",
+            "BroadcastName": "G RUSSELL",
+            "Abbreviation": "RUS",
+            "DriverId": "russell",
+            "TeamName": "Mercedes",
+            "TeamColor": "6CD3BF",
+            "TeamId": "mercedes",
+            "FirstName": "George",
+            "LastName": "Russell",
+            "FullName": "George Russell",
+            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/G/GEORUS01_George_Russell/georus01.png.transform/1col/image.png",
+            "CountryCode": "GBR",
+            "Position": 4,
+            "ClassifiedPosition": "4",
+            "GridPosition": 6,
+            "Q1": None,
+            "Q2": None,
+            "Q3": None,
+            "Time": 33229,
+            "Status": "Finished",
+            "Points": 12,
+        },
+        {
+            "DriverNumber": "55",
+            "BroadcastName": "C SAINZ",
+            "Abbreviation": "SAI",
+            "DriverId": "sainz",
+            "TeamName": "Ferrari",
+            "TeamColor": "F91536",
+            "TeamId": "ferrari",
+            "FirstName": "Carlos",
+            "LastName": "Sainz",
+            "FullName": "Carlos Sainz",
+            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/C/CARSAI01_Carlos_Sainz/carsai01.png.transform/1col/image.png",
+            "CountryCode": "ESP",
+            "Position": 5,
+            "ClassifiedPosition": "5",
+            "GridPosition": 3,
+            "Q1": None,
+            "Q2": None,
+            "Q3": None,
+            "Time": 42511,
+            "Status": "Finished",
+            "Points": 10,
+        },
+        {
+            "DriverNumber": "44",
+            "BroadcastName": "L HAMILTON",
+            "Abbreviation": "HAM",
+            "DriverId": "hamilton",
+            "TeamName": "Mercedes",
+            "TeamColor": "6CD3BF",
+            "TeamId": "mercedes",
+            "FirstName": "Lewis",
+            "LastName": "Hamilton",
+            "FullName": "Lewis Hamilton",
+            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/L/LEWHAM01_Lewis_Hamilton/lewham01.png.transform/1col/image.png",
+            "CountryCode": "GBR",
+            "Position": 6,
+            "ClassifiedPosition": "6",
+            "GridPosition": 13,
+            "Q1": None,
+            "Q2": None,
+            "Q3": None,
+            "Time": 51249,
+            "Status": "Finished",
+            "Points": 8,
+        },
+        {
+            "DriverNumber": "16",
+            "BroadcastName": "C LECLERC",
+            "Abbreviation": "LEC",
+            "DriverId": "leclerc",
+            "TeamName": "Ferrari",
+            "TeamColor": "F91536",
+            "TeamId": "ferrari",
+            "FirstName": "Charles",
+            "LastName": "Leclerc",
+            "FullName": "Charles Leclerc",
+            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/C/CHALEC01_Charles_Leclerc/chalec01.png.transform/1col/image.png",
+            "CountryCode": "MON",
+            "Position": 7,
+            "ClassifiedPosition": "7",
+            "GridPosition": 7,
+            "Q1": None,
+            "Q2": None,
+            "Q3": None,
+            "Time": 52988,
+            "Status": "Finished",
+            "Points": 6,
+        },
+        {
+            "DriverNumber": "10",
+            "BroadcastName": "P GASLY",
+            "Abbreviation": "GAS",
+            "DriverId": "gasly",
+            "TeamName": "Alpine",
+            "TeamColor": "2293D1",
+            "TeamId": "alpine",
+            "FirstName": "Pierre",
+            "LastName": "Gasly",
+            "FullName": "Pierre Gasly",
+            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/P/PIEGAS01_Pierre_Gasly/piegas01.png.transform/1col/image.png",
+            "CountryCode": "FRA",
+            "Position": 8,
+            "ClassifiedPosition": "8",
+            "GridPosition": 5,
+            "Q1": None,
+            "Q2": None,
+            "Q3": None,
+            "Time": 55670,
+            "Status": "Finished",
+            "Points": 4,
+        },
+        {
+            "DriverNumber": "31",
+            "BroadcastName": "E OCON",
+            "Abbreviation": "OCO",
+            "DriverId": "ocon",
+            "TeamName": "Alpine",
+            "TeamColor": "2293D1",
+            "TeamId": "alpine",
+            "FirstName": "Esteban",
+            "LastName": "Ocon",
+            "FullName": "Esteban Ocon",
+            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/E/ESTOCO01_Esteban_Ocon/estoco01.png.transform/1col/image.png",
+            "CountryCode": "FRA",
+            "Position": 9,
+            "ClassifiedPosition": "9",
+            "GridPosition": 8,
+            "Q1": None,
+            "Q2": None,
+            "Q3": None,
+            "Time": 58123,
+            "Status": "Finished",
+            "Points": 2,
+        },
+        {
+            "DriverNumber": "20",
+            "BroadcastName": "K MAGNUSSEN",
+            "Abbreviation": "MAG",
+            "DriverId": "kevin_magnussen",
+            "TeamName": "Haas F1 Team",
+            "TeamColor": "B6BABD",
+            "TeamId": "haas",
+            "FirstName": "Kevin",
+            "LastName": "Magnussen",
+            "FullName": "Kevin Magnussen",
+            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/K/KEVMAG01_Kevin_Magnussen/kevmag01.png.transform/1col/image.png",
+            "CountryCode": "DEN",
+            "Position": 10,
+            "ClassifiedPosition": "10",
+            "GridPosition": 4,
+            "Q1": None,
+            "Q2": None,
+            "Q3": None,
+            "Time": 62945,
+            "Status": "Finished",
+            "Points": 1,
+        },
+        {
+            "DriverNumber": "22",
+            "BroadcastName": "Y TSUNODA",
+            "Abbreviation": "TSU",
+            "DriverId": "tsunoda",
+            "TeamName": "AlphaTauri",
+            "TeamColor": "5E8FAA",
+            "TeamId": "alphatauri",
+            "FirstName": "Yuki",
+            "LastName": "Tsunoda",
+            "FullName": "Yuki Tsunoda",
+            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/Y/YUKTSU01_Yuki_Tsunoda/yuktsu01.png.transform/1col/image.png",
+            "CountryCode": "JPN",
+            "Position": 11,
+            "ClassifiedPosition": "11",
+            "GridPosition": 17,
+            "Q1": None,
+            "Q2": None,
+            "Q3": None,
+            "Time": 64309,
+            "Status": "Finished",
+            "Points": 0,
+        },
+        {
+            "DriverNumber": "18",
+            "BroadcastName": "L STROLL",
+            "Abbreviation": "STR",
+            "DriverId": "stroll",
+            "TeamName": "Aston Martin",
+            "TeamColor": "358C75",
+            "TeamId": "aston_martin",
+            "FirstName": "Lance",
+            "LastName": "Stroll",
+            "FullName": "Lance Stroll",
+            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/L/LANSTR01_Lance_Stroll/lanstr01.png.transform/1col/image.png",
+            "CountryCode": "CAN",
+            "Position": 12,
+            "ClassifiedPosition": "12",
+            "GridPosition": 18,
+            "Q1": None,
+            "Q2": None,
+            "Q3": None,
+            "Time": 64754,
+            "Status": "Finished",
+            "Points": 0,
+        },
+        {
+            "DriverNumber": "77",
+            "BroadcastName": "V BOTTAS",
+            "Abbreviation": "BOT",
+            "DriverId": "bottas",
+            "TeamName": "Alfa Romeo",
+            "TeamColor": "C92D4B",
+            "TeamId": "alfa",
+            "FirstName": "Valtteri",
+            "LastName": "Bottas",
+            "FullName": "Valtteri Bottas",
+            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/V/VALBOT01_Valtteri_Bottas/valbot01.png.transform/1col/image.png",
+            "CountryCode": "FIN",
+            "Position": 13,
+            "ClassifiedPosition": "13",
+            "GridPosition": 10,
+            "Q1": None,
+            "Q2": None,
+            "Q3": None,
+            "Time": 71637,
+            "Status": "Finished",
+            "Points": 0,
+        },
+        {
+            "DriverNumber": "23",
+            "BroadcastName": "A ALBON",
+            "Abbreviation": "ALB",
+            "DriverId": "albon",
+            "TeamName": "Williams",
+            "TeamColor": "37BEDD",
+            "TeamId": "williams",
+            "FirstName": "Alexander",
+            "LastName": "Albon",
+            "FullName": "Alexander Albon",
+            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/A/ALEALB01_Alexander_Albon/alealb01.png.transform/1col/image.png",
+            "CountryCode": "THA",
+            "Position": 14,
+            "ClassifiedPosition": "14",
+            "GridPosition": 11,
+            "Q1": None,
+            "Q2": None,
+            "Q3": None,
+            "Time": 72861,
+            "Status": "Finished",
+            "Points": 0,
+        },
+        {
+            "DriverNumber": "27",
+            "BroadcastName": "N HULKENBERG",
+            "Abbreviation": "HUL",
+            "DriverId": "hulkenberg",
+            "TeamName": "Haas F1 Team",
+            "TeamColor": "B6BABD",
+            "TeamId": "haas",
+            "FirstName": "Nico",
+            "LastName": "Hulkenberg",
+            "FullName": "Nico Hulkenberg",
+            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/N/NICHUL01_Nico_Hulkenberg/nichul01.png.transform/1col/image.png",
+            "CountryCode": "GER",
+            "Position": 15,
+            "ClassifiedPosition": "15",
+            "GridPosition": 12,
+            "Q1": None,
+            "Q2": None,
+            "Q3": None,
+            "Time": 74950,
+            "Status": "Finished",
+            "Points": 0,
+        },
+        {
+            "DriverNumber": "24",
+            "BroadcastName": "G ZHOU",
+            "Abbreviation": "ZHO",
+            "DriverId": "zhou",
+            "TeamName": "Alfa Romeo",
+            "TeamColor": "C92D4B",
+            "TeamId": "alfa",
+            "FirstName": "Guanyu",
+            "LastName": "Zhou",
+            "FullName": "Guanyu Zhou",
+            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/G/GUAZHO01_Guanyu_Zhou/guazho01.png.transform/1col/image.png",
+            "CountryCode": "CHN",
+            "Position": 16,
+            "ClassifiedPosition": "16",
+            "GridPosition": 14,
+            "Q1": None,
+            "Q2": None,
+            "Q3": None,
+            "Time": 78440,
+            "Status": "Finished",
+            "Points": 0,
+        },
+        {
+            "DriverNumber": "4",
+            "BroadcastName": "L NORRIS",
+            "Abbreviation": "NOR",
+            "DriverId": "norris",
+            "TeamName": "McLaren",
+            "TeamColor": "F58020",
+            "TeamId": "mclaren",
+            "FirstName": "Lando",
+            "LastName": "Norris",
+            "FullName": "Lando Norris",
+            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/L/LANNOR01_Lando_Norris/lannor01.png.transform/1col/image.png",
+            "CountryCode": "GBR",
+            "Position": 17,
+            "ClassifiedPosition": "17",
+            "GridPosition": 16,
+            "Q1": None,
+            "Q2": None,
+            "Q3": None,
+            "Time": 87717,
+            "Status": "Finished",
+            "Points": 0,
+        },
+        {
+            "DriverNumber": "21",
+            "BroadcastName": "N DE VRIES",
+            "Abbreviation": "DEV",
+            "DriverId": "de_vries",
+            "TeamName": "AlphaTauri",
+            "TeamColor": "5E8FAA",
+            "TeamId": "alphatauri",
+            "FirstName": "Nyck",
+            "LastName": "De Vries",
+            "FullName": "Nyck De Vries",
+            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/N/NYCDEV01_Nyck_De%20Vries/nycdev01.png.transform/1col/image.png",
+            "CountryCode": "NED",
+            "Position": 18,
+            "ClassifiedPosition": "18",
+            "GridPosition": 15,
+            "Q1": None,
+            "Q2": None,
+            "Q3": None,
+            "Time": 88949,
+            "Status": "Finished",
+            "Points": 0,
+        },
+        {
+            "DriverNumber": "81",
+            "BroadcastName": "O PIASTRI",
+            "Abbreviation": "PIA",
+            "DriverId": "piastri",
+            "TeamName": "McLaren",
+            "TeamColor": "F58020",
+            "TeamId": "mclaren",
+            "FirstName": "Oscar",
+            "LastName": "Piastri",
+            "FullName": "Oscar Piastri",
+            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/O/OSCPIA01_Oscar_Piastri/oscpia01.png.transform/1col/image.png",
+            "CountryCode": "AUS",
+            "Position": 19,
+            "ClassifiedPosition": "19",
+            "GridPosition": 19,
+            "Q1": None,
+            "Q2": None,
+            "Q3": None,
+            "Time": None,
+            "Status": "+1 Lap",
+            "Points": 0,
+        },
+        {
+            "DriverNumber": "2",
+            "BroadcastName": "L SARGEANT",
+            "Abbreviation": "SAR",
+            "DriverId": "sargeant",
+            "TeamName": "Williams",
+            "TeamColor": "37BEDD",
+            "TeamId": "williams",
+            "FirstName": "Logan",
+            "LastName": "Sargeant",
+            "FullName": "Logan Sargeant",
+            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/L/LOGSAR01_Logan_Sargeant/logsar01.png.transform/1col/image.png",
+            "CountryCode": "USA",
+            "Position": 20,
+            "ClassifiedPosition": "20",
+            "GridPosition": 20,
+            "Q1": None,
+            "Q2": None,
+            "Q3": None,
+            "Time": None,
+            "Status": "+1 Lap",
+            "Points": 0,
+        },
+    ]
+
+
+def test_get_results_with_session():
+    response = client.get("/results/2023/5?session=4")
+    assert response.status_code == status.HTTP_200_OK
+    assert response.json() == [
+        {
+            "DriverNumber": "11",
+            "BroadcastName": "S PEREZ",
+            "Abbreviation": "PER",
+            "DriverId": "perez",
+            "TeamName": "Red Bull Racing",
+            "TeamColor": "3671C6",
+            "TeamId": "red_bull",
+            "FirstName": "Sergio",
+            "LastName": "Perez",
+            "FullName": "Sergio Perez",
+            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/S/SERPER01_Sergio_Perez/serper01.png.transform/1col/image.png",
+            "CountryCode": "MEX",
+            "Position": 1,
+            "ClassifiedPosition": "",
+            "GridPosition": None,
+            "Q1": 87713,
+            "Q2": 87328,
+            "Q3": 86841,
+            "Time": None,
+            "Status": "",
+            "Points": None,
+        },
+        {
+            "DriverNumber": "14",
+            "BroadcastName": "F ALONSO",
+            "Abbreviation": "ALO",
+            "DriverId": "alonso",
+            "TeamName": "Aston Martin",
+            "TeamColor": "358C75",
+            "TeamId": "aston_martin",
+            "FirstName": "Fernando",
+            "LastName": "Alonso",
+            "FullName": "Fernando Alonso",
+            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/F/FERALO01_Fernando_Alonso/feralo01.png.transform/1col/image.png",
+            "CountryCode": "ESP",
+            "Position": 2,
+            "ClassifiedPosition": "",
+            "GridPosition": None,
+            "Q1": 88179,
+            "Q2": 87097,
+            "Q3": 87202,
+            "Time": None,
+            "Status": "",
+            "Points": None,
+        },
+        {
+            "DriverNumber": "55",
+            "BroadcastName": "C SAINZ",
+            "Abbreviation": "SAI",
+            "DriverId": "sainz",
+            "TeamName": "Ferrari",
+            "TeamColor": "F91536",
+            "TeamId": "ferrari",
+            "FirstName": "Carlos",
+            "LastName": "Sainz",
+            "FullName": "Carlos Sainz",
+            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/C/CARSAI01_Carlos_Sainz/carsai01.png.transform/1col/image.png",
+            "CountryCode": "ESP",
+            "Position": 3,
+            "ClassifiedPosition": "",
+            "GridPosition": None,
+            "Q1": 87686,
+            "Q2": 87148,
+            "Q3": 87349,
+            "Time": None,
+            "Status": "",
+            "Points": None,
+        },
+        {
+            "DriverNumber": "20",
+            "BroadcastName": "K MAGNUSSEN",
+            "Abbreviation": "MAG",
+            "DriverId": "kevin_magnussen",
+            "TeamName": "Haas F1 Team",
+            "TeamColor": "B6BABD",
+            "TeamId": "haas",
+            "FirstName": "Kevin",
+            "LastName": "Magnussen",
+            "FullName": "Kevin Magnussen",
+            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/K/KEVMAG01_Kevin_Magnussen/kevmag01.png.transform/1col/image.png",
+            "CountryCode": "DEN",
+            "Position": 4,
+            "ClassifiedPosition": "",
+            "GridPosition": None,
+            "Q1": 87809,
+            "Q2": 87673,
+            "Q3": 87767,
+            "Time": None,
+            "Status": "",
+            "Points": None,
+        },
+        {
+            "DriverNumber": "10",
+            "BroadcastName": "P GASLY",
+            "Abbreviation": "GAS",
+            "DriverId": "gasly",
+            "TeamName": "Alpine",
+            "TeamColor": "2293D1",
+            "TeamId": "alpine",
+            "FirstName": "Pierre",
+            "LastName": "Gasly",
+            "FullName": "Pierre Gasly",
+            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/P/PIEGAS01_Pierre_Gasly/piegas01.png.transform/1col/image.png",
+            "CountryCode": "FRA",
+            "Position": 5,
+            "ClassifiedPosition": "",
+            "GridPosition": None,
+            "Q1": 88061,
+            "Q2": 87612,
+            "Q3": 87786,
+            "Time": None,
+            "Status": "",
+            "Points": None,
+        },
+        {
+            "DriverNumber": "63",
+            "BroadcastName": "G RUSSELL",
+            "Abbreviation": "RUS",
+            "DriverId": "russell",
+            "TeamName": "Mercedes",
+            "TeamColor": "6CD3BF",
+            "TeamId": "mercedes",
+            "FirstName": "George",
+            "LastName": "Russell",
+            "FullName": "George Russell",
+            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/G/GEORUS01_George_Russell/georus01.png.transform/1col/image.png",
+            "CountryCode": "GBR",
+            "Position": 6,
+            "ClassifiedPosition": "",
+            "GridPosition": None,
+            "Q1": 88086,
+            "Q2": 87743,
+            "Q3": 87804,
+            "Time": None,
+            "Status": "",
+            "Points": None,
+        },
+        {
+            "DriverNumber": "16",
+            "BroadcastName": "C LECLERC",
+            "Abbreviation": "LEC",
+            "DriverId": "leclerc",
+            "TeamName": "Ferrari",
+            "TeamColor": "F91536",
+            "TeamId": "ferrari",
+            "FirstName": "Charles",
+            "LastName": "Leclerc",
+            "FullName": "Charles Leclerc",
+            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/C/CHALEC01_Charles_Leclerc/chalec01.png.transform/1col/image.png",
+            "CountryCode": "MON",
+            "Position": 7,
+            "ClassifiedPosition": "",
+            "GridPosition": None,
+            "Q1": 87713,
+            "Q2": 86964,
+            "Q3": 87861,
+            "Time": None,
+            "Status": "",
+            "Points": None,
+        },
+        {
+            "DriverNumber": "31",
+            "BroadcastName": "E OCON",
+            "Abbreviation": "OCO",
+            "DriverId": "ocon",
+            "TeamName": "Alpine",
+            "TeamColor": "2293D1",
+            "TeamId": "alpine",
+            "FirstName": "Esteban",
+            "LastName": "Ocon",
+            "FullName": "Esteban Ocon",
+            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/E/ESTOCO01_Esteban_Ocon/estoco01.png.transform/1col/image.png",
+            "CountryCode": "FRA",
+            "Position": 8,
+            "ClassifiedPosition": "",
+            "GridPosition": None,
+            "Q1": 87872,
+            "Q2": 87444,
+            "Q3": 87935,
+            "Time": None,
+            "Status": "",
+            "Points": None,
+        },
+        {
+            "DriverNumber": "1",
+            "BroadcastName": "M VERSTAPPEN",
+            "Abbreviation": "VER",
+            "DriverId": "max_verstappen",
+            "TeamName": "Red Bull Racing",
+            "TeamColor": "3671C6",
+            "TeamId": "red_bull",
+            "FirstName": "Max",
+            "LastName": "Verstappen",
+            "FullName": "Max Verstappen",
+            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/M/MAXVER01_Max_Verstappen/maxver01.png.transform/1col/image.png",
+            "CountryCode": "NED",
+            "Position": 9,
+            "ClassifiedPosition": "",
+            "GridPosition": None,
+            "Q1": 87363,
+            "Q2": 86814,
+            "Q3": None,
+            "Time": None,
+            "Status": "",
+            "Points": None,
+        },
+        {
+            "DriverNumber": "77",
+            "BroadcastName": "V BOTTAS",
+            "Abbreviation": "BOT",
+            "DriverId": "bottas",
+            "TeamName": "Alfa Romeo",
+            "TeamColor": "C92D4B",
+            "TeamId": "alfa",
+            "FirstName": "Valtteri",
+            "LastName": "Bottas",
+            "FullName": "Valtteri Bottas",
+            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/V/VALBOT01_Valtteri_Bottas/valbot01.png.transform/1col/image.png",
+            "CountryCode": "FIN",
+            "Position": 10,
+            "ClassifiedPosition": "",
+            "GridPosition": None,
+            "Q1": 87864,
+            "Q2": 87564,
+            "Q3": None,
+            "Time": None,
+            "Status": "",
+            "Points": None,
+        },
+        {
+            "DriverNumber": "23",
+            "BroadcastName": "A ALBON",
+            "Abbreviation": "ALB",
+            "DriverId": "albon",
+            "TeamName": "Williams",
+            "TeamColor": "37BEDD",
+            "TeamId": "williams",
+            "FirstName": "Alexander",
+            "LastName": "Albon",
+            "FullName": "Alexander Albon",
+            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/A/ALEALB01_Alexander_Albon/alealb01.png.transform/1col/image.png",
+            "CountryCode": "THA",
+            "Position": 11,
+            "ClassifiedPosition": "",
+            "GridPosition": None,
+            "Q1": 88234,
+            "Q2": 87795,
+            "Q3": None,
+            "Time": None,
+            "Status": "",
+            "Points": None,
+        },
+        {
+            "DriverNumber": "27",
+            "BroadcastName": "N HULKENBERG",
+            "Abbreviation": "HUL",
+            "DriverId": "hulkenberg",
+            "TeamName": "Haas F1 Team",
+            "TeamColor": "B6BABD",
+            "TeamId": "haas",
+            "FirstName": "Nico",
+            "LastName": "Hulkenberg",
+            "FullName": "Nico Hulkenberg",
+            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/N/NICHUL01_Nico_Hulkenberg/nichul01.png.transform/1col/image.png",
+            "CountryCode": "GER",
+            "Position": 12,
+            "ClassifiedPosition": "",
+            "GridPosition": None,
+            "Q1": 87945,
+            "Q2": 87903,
+            "Q3": None,
+            "Time": None,
+            "Status": "",
+            "Points": None,
+        },
+        {
+            "DriverNumber": "44",
+            "BroadcastName": "L HAMILTON",
+            "Abbreviation": "HAM",
+            "DriverId": "hamilton",
+            "TeamName": "Mercedes",
+            "TeamColor": "6CD3BF",
+            "TeamId": "mercedes",
+            "FirstName": "Lewis",
+            "LastName": "Hamilton",
+            "FullName": "Lewis Hamilton",
+            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/L/LEWHAM01_Lewis_Hamilton/lewham01.png.transform/1col/image.png",
+            "CountryCode": "GBR",
+            "Position": 13,
+            "ClassifiedPosition": "",
+            "GridPosition": None,
+            "Q1": 87846,
+            "Q2": 87975,
+            "Q3": None,
+            "Time": None,
+            "Status": "",
+            "Points": None,
+        },
+        {
+            "DriverNumber": "24",
+            "BroadcastName": "G ZHOU",
+            "Abbreviation": "ZHO",
+            "DriverId": "zhou",
+            "TeamName": "Alfa Romeo",
+            "TeamColor": "C92D4B",
+            "TeamId": "alfa",
+            "FirstName": "Guanyu",
+            "LastName": "Zhou",
+            "FullName": "Guanyu Zhou",
+            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/G/GUAZHO01_Guanyu_Zhou/guazho01.png.transform/1col/image.png",
+            "CountryCode": "CHN",
+            "Position": 14,
+            "ClassifiedPosition": "",
+            "GridPosition": None,
+            "Q1": 88180,
+            "Q2": 88091,
+            "Q3": None,
+            "Time": None,
+            "Status": "",
+            "Points": None,
+        },
+        {
+            "DriverNumber": "21",
+            "BroadcastName": "N DE VRIES",
+            "Abbreviation": "DEV",
+            "DriverId": "de_vries",
+            "TeamName": "AlphaTauri",
+            "TeamColor": "5E8FAA",
+            "TeamId": "alphatauri",
+            "FirstName": "Nyck",
+            "LastName": "De Vries",
+            "FullName": "Nyck De Vries",
+            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/N/NYCDEV01_Nyck_De%20Vries/nycdev01.png.transform/1col/image.png",
+            "CountryCode": "NED",
+            "Position": 15,
+            "ClassifiedPosition": "",
+            "GridPosition": None,
+            "Q1": 88325,
+            "Q2": 88395,
+            "Q3": None,
+            "Time": None,
+            "Status": "",
+            "Points": None,
+        },
+        {
+            "DriverNumber": "4",
+            "BroadcastName": "L NORRIS",
+            "Abbreviation": "NOR",
+            "DriverId": "norris",
+            "TeamName": "McLaren",
+            "TeamColor": "F58020",
+            "TeamId": "mclaren",
+            "FirstName": "Lando",
+            "LastName": "Norris",
+            "FullName": "Lando Norris",
+            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/L/LANNOR01_Lando_Norris/lannor01.png.transform/1col/image.png",
+            "CountryCode": "GBR",
+            "Position": 16,
+            "ClassifiedPosition": "",
+            "GridPosition": None,
+            "Q1": 88394,
+            "Q2": None,
+            "Q3": None,
+            "Time": None,
+            "Status": "",
+            "Points": None,
+        },
+        {
+            "DriverNumber": "22",
+            "BroadcastName": "Y TSUNODA",
+            "Abbreviation": "TSU",
+            "DriverId": "tsunoda",
+            "TeamName": "AlphaTauri",
+            "TeamColor": "5E8FAA",
+            "TeamId": "alphatauri",
+            "FirstName": "Yuki",
+            "LastName": "Tsunoda",
+            "FullName": "Yuki Tsunoda",
+            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/Y/YUKTSU01_Yuki_Tsunoda/yuktsu01.png.transform/1col/image.png",
+            "CountryCode": "JPN",
+            "Position": 17,
+            "ClassifiedPosition": "",
+            "GridPosition": None,
+            "Q1": 88429,
+            "Q2": None,
+            "Q3": None,
+            "Time": None,
+            "Status": "",
+            "Points": None,
+        },
+        {
+            "DriverNumber": "18",
+            "BroadcastName": "L STROLL",
+            "Abbreviation": "STR",
+            "DriverId": "stroll",
+            "TeamName": "Aston Martin",
+            "TeamColor": "358C75",
+            "TeamId": "aston_martin",
+            "FirstName": "Lance",
+            "LastName": "Stroll",
+            "FullName": "Lance Stroll",
+            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/L/LANSTR01_Lance_Stroll/lanstr01.png.transform/1col/image.png",
+            "CountryCode": "CAN",
+            "Position": 18,
+            "ClassifiedPosition": "",
+            "GridPosition": None,
+            "Q1": 88476,
+            "Q2": None,
+            "Q3": None,
+            "Time": None,
+            "Status": "",
+            "Points": None,
+        },
+        {
+            "DriverNumber": "81",
+            "BroadcastName": "O PIASTRI",
+            "Abbreviation": "PIA",
+            "DriverId": "piastri",
+            "TeamName": "McLaren",
+            "TeamColor": "F58020",
+            "TeamId": "mclaren",
+            "FirstName": "Oscar",
+            "LastName": "Piastri",
+            "FullName": "Oscar Piastri",
+            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/O/OSCPIA01_Oscar_Piastri/oscpia01.png.transform/1col/image.png",
+            "CountryCode": "AUS",
+            "Position": 19,
+            "ClassifiedPosition": "",
+            "GridPosition": None,
+            "Q1": 88484,
+            "Q2": None,
+            "Q3": None,
+            "Time": None,
+            "Status": "",
+            "Points": None,
+        },
+        {
+            "DriverNumber": "2",
+            "BroadcastName": "L SARGEANT",
+            "Abbreviation": "SAR",
+            "DriverId": "sargeant",
+            "TeamName": "Williams",
+            "TeamColor": "37BEDD",
+            "TeamId": "williams",
+            "FirstName": "Logan",
+            "LastName": "Sargeant",
+            "FullName": "Logan Sargeant",
+            "HeadshotUrl": "https://www.formula1.com/content/dam/fom-website/drivers/L/LOGSAR01_Logan_Sargeant/logsar01.png.transform/1col/image.png",
+            "CountryCode": "USA",
+            "Position": 20,
+            "ClassifiedPosition": "",
+            "GridPosition": None,
+            "Q1": 88577,
+            "Q2": None,
+            "Q3": None,
+            "Time": None,
+            "Status": "",
+            "Points": None,
+        },
+    ]
+
+
+# endregion results - good inputs
+
+# region results - bad inputs
+
+
+def test_get_results_bad_round_invalid():
+    response = client.get("/results/2023/25?session=2")
+    assert response.status_code == status.HTTP_400_BAD_REQUEST
+    assert response.json() == {"detail": "Bad Request. Invalid round: 25"}
+
+
+# endregion results - bad inputs
+
+# endregion results
